% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_confirm.R
\name{bayes_confirm}
\alias{bayes_confirm}
\title{Title}
\usage{
bayes_confirm(X, nu, delta, type, hyp = NULL, n_samples, cores)
}
\arguments{
\item{X}{}

\item{nu}{}

\item{delta}{}

\item{type}{}

\item{hyp}{}

\item{n_samples}{}

\item{cores}{}
}
\description{
Title
}
