% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_explore.R
\name{bayes_explore}
\alias{bayes_explore}
\title{bayes_explore}
\usage{
bayes_explore(X, exhaustive = FALSE, adjustment = NULL, sample = NULL,
  prior = NULL)
}
\arguments{
\item{X}{}

\item{exhaustive}{}

\item{adjustment}{}

\item{sample}{}

\item{prior}{}
}
\description{
bayes_explore
}
