% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_R2.R
\name{bayes_R2}
\alias{bayes_R2}
\title{Bayesian R2}
\usage{
bayes_R2(fit, newdata = NULL, selected, ci_width = 0.95, samples = 500)
}
\arguments{
\item{fit}{object of class bayes_estimate}

\item{newdata}{out-of-sample option}

\item{selected}{the selected adjacency matrix}

\item{ci_width}{credible interval width for Bayes R2}

\item{samples}{Number of samples for compute Bayes R2}
}
\description{
Bayesian R2
}
\examples{
X <- BGGM::ptsd

# X is the data matrix of dimensions n by p
fit <- bayes_estimate(X)

# select the graphical structure
select <- estimate_select(fit, ci_width = .95)

# compute Bayesian R2
r2 <-  bayes_R2(fit, selected = select$adjacency_mat, ci_width = .95, samples = 1000)
}
