% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_summary.R
\name{beta_summary}
\alias{beta_summary}
\title{Title}
\usage{
beta_summary(x, node, ci_width, samples)
}
\arguments{
\item{x}{object \code{bayes_estimate}}

\item{node}{which nodes to summarize}

\item{ci_width}{credible interval}

\item{samples}{number of posterior samples}
}
\description{
Title
}
