% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bfi}
\alias{bfi}
\title{25 Personality items representing 5 factors}
\format{A data frame with 25 variables and 2800 observations (but with missing data)}
\description{
This data set and the documentation was taken from the \strong{psych} package. Further details can be found in the documentation
of the \strong{psych} package.
}
\details{
\itemize{
  \item \code{A1} Am indifferent to the feelings of others. (q_146)
  \item \code{A2} Inquire about others' well-being. (q_1162)
  \item \code{A3} Know how to comfort others. (q_1206)
  \item \code{A4} Love children. (q_1364)
  \item \code{A5} Make people feel at ease. (q_1419)
  \item \code{C1} Am exacting in my work. (q_124)
  \item \code{C2} Continue until everything is perfect. (q_530)
  \item \code{C3} Do things according to a plan. (q_619)
  \item \code{C4} Do things in a half-way manner. (q_626)
  \item \code{C5} Waste my time. (q_1949)
  \item \code{E1} Don't talk a lot. (q_712)
  \item \code{E2} Find it difficult to approach others. (q_901)
  \item \code{E3} Know how to captivate people. (q_1205)
  \item \code{E4} Make friends easily. (q_1410)
  \item \code{E5} Take charge. (q_1768)
  \item \code{N1} Get angry easily. (q_952)
  \item \code{N2} Get irritated easily. (q_974)
  \item \code{N3} Have frequent mood swings. (q_1099)
  \item \code{N4} Often feel blue. (q_1479)
  \item \code{N5} Panic easily. (q_1505)
  \item \code{o1} Am full of ideas. (q_128)
  \item \code{o2} Avoid difficult reading material.(q_316)
  \item \code{o3} Carry the conversation to a higher level. (q_492)
  \item \code{o4} Spend time reflecting on things. (q_1738)
  \item \code{o5} Will not probe deeply into a subject. (q_1964)
  \item \code{gender} Males = 1, Females =2
  \item \code{education} 1 = HS, 2 = finished HS, 3 = some college, 4 = college graduate 5 = graduate degree
}
}
\references{
Revelle, W. (2018) psych: Procedures for Personality and Psychological Research, Northwestern University,
Evanston, Illinois, USA, https://CRAN.R-project.org/package=psych Version = 1.8.12.
}
\keyword{datasets}
