% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGGM.R
\name{BGGM}
\alias{BGGM}
\title{Title}
\usage{
BGGM(x, chains = 2, iter = 1000, burnin = 100)
}
\arguments{
\item{x}{data matrix}

\item{chains}{number of chains}

\item{iter}{number of posterior samples for each chain}

\item{burnin}{discarded samples during warmup}
}
\description{
Title
}
