% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGGM_select.R
\name{BGGM_select}
\alias{BGGM_select}
\title{Title}
\usage{
BGGM_select(x, BF_directional = NULL, probability = NULL)
}
\arguments{
\item{x}{BGGM fitted object}

\item{BF_directional}{Decision rule based on Bayes factors}

\item{probability}{Decision rule based on Posterior probabiities}
}
\description{
Title
}
