% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{centrality}
\alias{centrality}
\title{Bayesian Centrality}
\usage{
centrality(
  object,
  type = "strength",
  cluster = NULL,
  select = FALSE,
  cred = NULL,
  BF_cut = NULL
)
}
\arguments{
\item{object}{an object of class \code{estimate} or \code{explore}. All data types (e.g., ordinal, binary, etc.)
are supported.}

\item{type}{character string. Which type of centrality? Currently the only options are
\code{strength} and \code{bridge_strength}. See note for further details.}

\item{cluster}{charcter string. Which clusters or communities do the nodes belong to? This is
required when \code{type = bridge_stength}.}

\item{select}{logical. Should strength be condtional on the selected model (default set to \code{FALSE})? See note for further details.}

\item{cred}{numeric. Credible interval width used for selecting the graph with an \code{estimate}
object (default is \code{0.95}).}

\item{BF_cut}{numeric. Evidentiary threshold used for selected the graph with an \code{explore}
object (default is \code{3})}
}
\value{
object of class \code{BGGM} and \code{centrality}
}
\description{
Compute Bayesian centrality. Currently the only option is for strength (see note for details).
For each variable, this provides a posterior distribution for the sum of the respective partial
correlations. Testing strength (e.g., differences or confirmatory testing of order contraints) is
implemented in \code{test.centrality}.
}
\note{
Centrality is a new feature to \strong{BGGM} (> 1.0). At this time, only centrality indices based
on the partial correlations (e.g., strength) are implemented and this will not likely change
\insertCite{@see critiques on centrality in @bringmann2019centrality; @hallquist2019problems}{BGGM}. This is because
strength is \emph{merely} the sum of partial correlations, which have a known (approximate) distribution
(\href{https://en.wikipedia.org/wiki/Partial_correlation}{Wikipedia}). This
translates into the sum also having those same properties. This is ideal for making statistical inference. In fact,
it is not uncommon to test the sum of coefficients with classical methods
\insertCite{@page 96, section 4.5.2, in  @baum2006introduction;textual}{BGGM}.

We are open to all suggestions for implementing strength related measures: perhaps there is a particular sum
of theorectial and/or substantative interest.

\strong{Beware of Model Selection Bias:}

The default setting is \code{select = FALSE}. This means that \emph{all} of the partial correlation for a given
node are summed and not only those that were determined to be non-zero. Although it may seem that including only
the selected edges leads to desirable "inference," there is an important caveat to consider.
This induces a problem known as model selection bias that arises when \emph{conditioning} on the selected model..

To make sense of this, note that:

\enumerate{
\item To be selected, the partial correlations must far enough way from zero.
\item As a result of 1, edges will often be upward biased becauase they are only considered
      when they are selected \insertCite{@more specifically a truncated sampling distribution, p. 5 in  @leeb2006can}{BGGM}
\item This can translate into strength also having issues when computing it from only the selected edges.
}

Overcoming this issue is often called 'selective' or 'post-selection' inference
(see the references in \href{http://bactra.org/notebooks/post-model-selection-inference.html}{weblink}).
A viable and simple approach is to \strong{not condition} strength on the selected model
(i.e., \code{select = FALSE}). This readily allows for comparing the sum of partial correlations
(e.g., differences in bridge strength). In future versions, the option \code{select = FALSE} may be removed
altogether due to the innate proclivity to compare estimates visually (plot.centrality). These implicit
comparisons should directly be tested with (see \code{test.centrality}) that  requires \code{select = FALSE}.
}
\references{
\insertAllCited{}
}
