% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.estimate.R
\name{coef.estimate}
\alias{coef.estimate}
\title{Precision Matrix to Multiple Regression}
\usage{
\method{coef}{estimate}(object, node = 1, CrI = 0.95, iter = 1000,
  ...)
}
\arguments{
\item{node}{which node to summarize (i.e., the outcome)}

\item{CrI}{credible interval used in the summary output}

\item{iter}{number of samples used in the conversion.}

\item{...}{e.g., \code{digits}}

\item{fit}{object of class \code{estimate} (analytic = F)}
}
\value{
summary of the multiple regression coefficients for \code{node}
}
\description{
There is a direct correspondence between the covariance matrix and multiple regression. In the case of GGMs, it is possible
to estimate the edge set with multiple regression (i.e., neighborhood selection). In *BGGM*, the precision matrix is first sampled from, and then
each draws is converted to the corresponding coefficients and error variances. This results in a posterior distribution. This function can be used
to perform Bayesian multiple regression.
}
\examples{

# p = 10
Y <- BGGM::bfi[,1:10]

# sample posterior
fit <- estimate(Y, iter = 5000)

# precision to regression
coefficients(fit, node = 1, ci_width = 0.95)
}
