% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.estimate.R
\name{coef.explore}
\alias{coef.explore}
\title{Compute Regression Parameters for \code{explore} Objects}
\usage{
\method{coef}{explore}(object, iter = NULL, ...)
}
\arguments{
\item{object}{object of class \code{explore}}

\item{iter}{number of posterior samples (defaults to the number in the object).}

\item{...}{not currently used.}
}
\value{

}
\description{
Compute Regression Parameters for \code{explore} Objects
}
\examples{
# p = 10
Y <- BGGM::bfi[,1:10]

# sample posterior
fit <- estimate(Y, iter = 5000)

# precision to regression
coefficients(fit, node = 1, cred = 0.95)
}
