% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.estimate.R, R/edge_compare.estimate.R
\name{compare.estimate}
\alias{compare.estimate}
\title{Title}
\usage{
\method{compare}{estimate}(x, contrast, ci_width, rope = NULL)

\method{compare}{estimate}(x, contrast, ci_width, rope = NULL)
}
\arguments{
\item{rope}{}

\item{mute}{}
}
\description{
Title

Title
}
