% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirm.default.R
\name{confirm.default}
\alias{confirm.default}
\title{Confirmatory Hypothesis Testing}
\usage{
\method{confirm}{default}(x, hypothesis, prior_sd, iter = 25000,
  cores = 2)
}
\arguments{
\item{x}{data matrix (\emph{n} \times  \emph{p})}

\item{hypothesis}{hypotheses to be compared}

\item{prior_sd}{hypothesized standard deviation for $\rho$}

\item{iter}{posterior and prior samples. 25,000 is the default, as it results in a more stable Bayes factor than using, say, 5,000.}

\item{cores}{number of cores for parallel computing. The defaul is 2, but this can be adjusted}
}
\value{
An object of class \code{confirm}
}
\description{
Confirmatory Hypothesis Testing
}
\note{
Currently inequality and equality restrictions can be tested. The former is an ordering the respective edge sizes,
whereas the latter allows for testing whether certian edges are exactly the same.
}
\examples{


Y <- BGGM::bfi[,1:5]

hypothesis <- c("3--5 > 1--5 > 2--5;
                3--5 > 2--5 > 1--5")

                test_order <-  confirm(x = x, hypothesis  = hypothesis,
                prior_sd = 0.5, iter = 50000,
                cores = 2)
}
