% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_compare.explore.R
\name{edge_compare.explore}
\alias{edge_compare.explore}
\title{Edge (Partial Correaltion) Contrasts with the Bayes Factor}
\usage{
\method{edge_compare}{explore}(x, contrast, alternative)
}
\arguments{
\item{x}{object from \code{explore}}

\item{contrast}{partial correlations to compare}

\item{alternative}{\code{greater}, \code{less}, \code{two.sided}, or \code{exhaustive}}
}
\value{
list of class \code{edge_compare.explore}:
\itemize{
\item \code{results} data.frame summary of the edge contrasts
\item \code{alternative} \code{greater}, \code{less}, \code{two.sided}, or \code{exhaustive}
\item \code{call} match.call()
}
}
\description{
Allows for comparing partial correlations \emph{within} the same GGM--e.g., to determine the largest edge for each node. The differences are tested
with the Bayes factor. The test can be either directional, for example whether one edge is larger or smaller than another, or two sided testing is also possible. Further,
there is also an exhaustive option that provides the posterior probability of the null, greater than zero, and less than zero.
}
\examples{

# p = 10
Y <- BGGM::bfi[,1:10]

# sample from posterior
fit_bf <- explore(Y, prior_sd = 0.5,
              iter = 5000,
              cores = 2)

# edge compare
 edge_comp <- edge_compare(fit_bf,
                         contrast = c("1--5 - 1--3",
                                      "1--2 - 1--6",
                                      "1--4 - 1--7",
                                      "1--5 - 1--10",
                                      "1--2 - 1--9"),
                         alternative = "exhaustive")
# summary
summary(edge_comp)
}
