% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate.default}
\alias{estimate.default}
\title{Gaussian Graphical Models with Credible Intervals or the Region of Practical Equivalence}
\usage{
\method{estimate}{default}(x, samples = 5000, analytic = FALSE)
}
\arguments{
\item{x}{data matrix (\emph{n} $\times$  \emph{p}).}

\item{samples}{number of posterior samples}

\item{analytic}{analytic solution. see notes for futher details.}
}
\value{
An object of class \code{estimate}.
}
\description{
Learn the conditional (in)dependence structure with credible intervals or the region of practical equivalence.
For the former, there is an analytic solution avaiable, whereas for the latter, samples are efficiently drawn from the posterior
distribution.
}
\note{
The default is to draws samples from the posterior distribution (\code{analytic = FALSE}). The samples are required for computing edge differences,
Bayesian R2, etc. If the goal is to *only* determined the non-zero effects, this can be accomplished by setting \code{analytic = TRUE}. This is accomplished
by estimating the posterior mean and variance, from which the credible intervals can computed. Note also sampling is very fast--i.e., less than 1 second
with p = 25, n = 2500 and 5,000 samples. There is one function that makes use of the analytic solution. Namely, \code{loocv} computes node-wise leave-one-out
error (also analytically).
}
\examples{

# p = 5
Y <- BGGM::bfi[,1:20]

# analytic approach (sample by setting analytic = F)
fit_analytic <- estimate(Y, analytic = T)

# select the graph (edge set E)
E <- select(fit_analytic, ci_width = 0.95)

# plot
plot(E, type = "network")
}
