% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{GGM: Estimation}
\usage{
estimate(
  Y,
  formula = NULL,
  type = "continuous",
  mixed_type = NULL,
  analytic = FALSE,
  prior_sd = 0.25,
  iter = 5000,
  ...
)
}
\arguments{
\item{Y}{matrix (or data frame) of dimensions \emph{n} (observations) by  \emph{p} (variables).}

\item{formula}{an object of class \code{\link[stats]{formula}}. This allows for including
control variables in the model (i.e., \code{~ gender}). See the note for further details.}

\item{type}{character string. Which type of data for \strong{Y} ? The options include \code{continuous},
\code{binary}, \code{ordinal}, or \code{mixed}. See the note for further details.}

\item{mixed_type}{numeric vector. An indicator of length p for which varibles should be treated as ranks.
(1 for rank and 0 to assume normality). The default is currently to treat all integer variables as ranks
when \code{type = "mixed"} and \code{NULL} otherwise. See note for further details.

Treating continous data as ranks is computationally expensive and can be avoided by assuming normality.}

\item{analytic}{logical. Should the analytic solution be computed (default is \code{FALSE})?}

\item{iter}{number of iterations (posterior samples; defaults to 5000).}

\item{...}{currently ignored.}
}
\value{
list of class \code{estimate}:

\code{analytic = TRUE}:
\itemize{
\item \code{fit} list of analytic solution estimates
\itemize{
\item \code{inv_mu} inverse covariance matrix (mean)
\item \code{inv_var} inverse covariance matrix (variance)
\item \code{partial} partial correlation matrix
}
\item \code{analytic} TRUE
\item \code{call} match.call()
\item \code{dat} data matrix
\item \code{p} number of variables
}

\code{analytic = FALSE}:
\itemize{
\item \code{parcors_mat} partial correlation matrix
\item \code{inv_mat} inverse covariance matrix
\item \code{posterior samples} posterior samples for partial correlations and inverse covariance matrix
\item \code{p} number of variables
\item \code{dat} data matrix
\item \code{iter} number of posterior samples
\item \code{call} match.call()
\item \code{analytic} FALSE
}
}
\description{
Estimate the conditional (in)dependence with either an analytic solution or efficiently
sampling from the posterior distribution. These methods were introduced in \insertCite{Williams2019;textual}{BGGM}.
The graph is then selected with \code{\link{select.estimate}}, with either directional posterior probabilities
\insertCite{Marsman2017a}{BGGM}, credible intervals, or a region of practical equivalence \insertCite{Kruschke2017}{BGGM}.
Bayesian hypothesis testing is implemented in \code{\link{explore}} and \code{\link{confirm}} \insertCite{Williams2019_bf}{BGGM}.
}
\note{
The default is to draw samples from the posterior distribution (\code{analytic = FALSE}). The samples are
required for computing edge differences (see \code{\link{ggm_compare_estimate}}), Bayesian R2 introduced in
\insertCite{gelman_r2_2019;textual}{BGGM} (see \code{\link{bayes_R2}}), etc. If the goal is to *only* determine
the non-zero effects, this can be accomplished by setting \code{analytic = TRUE}.


\strong{Controlling for Variables}:

When controlling for variables, it is assumed that \code{Y} includes \emph{only}
the nodes in the GGM and the control variables. Internally, \code{only} the predictors
that are included in \code{formula} are removed from \code{Y}. An example is provided below.

\strong{Dealing with Errors}:

An error is most likely to arise when \code{type = "ordinal"}. The are two common errors (although still rare).
The first is due to sampling the thresholds, especially when the data is
heavily skewed. This can result in an ill-defined matrix. If this occurs, we recommend to first try
decreasing \code{prior_sd} (i.e., a more informative prior). If that does not work, then  change the
data type to \code{type = mixed} which estimates a copula GGM. This should work without a problem.

The second common error is from the categories. For example, if the error staes that the index is out of bounds, this
means there are zeros in the data. The first category must be \code{1}. This can easily be addressed by adding 1 to the
data matrix.


\strong{Interpretation of conditional (in)dependence models for latent data}:

See \code{\link{BGGM-package}} for details about interpreting GGMs based on latent data
(i.e, all data types besides \code{"continuous"})
}
\examples{
# p = 5
Y <- BGGM::bfi[, 1:5]

# analytic approach (sample by setting analytic = FALSE)
fit_analytic <- estimate(Y, analytic = TRUE)

# select the graph (edge set E)
E <- select(fit_analytic, ci_width = 0.95)

}
\references{
\insertAllCited{}
}
