% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_z2r.R
\name{fisher_z_to_r}
\alias{fisher_z_to_r}
\title{Fisher Z Back Transformation}
\usage{
fisher_z_to_r(z)
}
\arguments{
\item{z}{Fisher Z}
}
\value{
Correlation (s) (backtransformed)
}
\description{
Back tranform Fisher's Z to correlations
}
\examples{
fisher_z_to_r(0.5)
}
