% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_ordinal.R
\name{gen_ordinal}
\alias{gen_ordinal}
\title{Generate Ordinal and Binary data}
\usage{
gen_ordinal(n, p, levels, cor_mat, empirical = FALSE)
}
\arguments{
\item{n}{number of observations}

\item{p}{number of variables}

\item{levels}{number of categories}

\item{cor_mat}{true correlation matrix}

\item{empirical}{logical. If true, mu and Sigma specify the empirical not
population mean and covariance matrix.}
}
\value{

}
\description{
Generate Ordinal and Binary data
}
\examples{

main <- BGGM::ptsd_cor1[1:5,1:5]
p <- ncol(main)

pcors <- -(cov2cor(solve(main)) -diag(p))
diag(pcors) <- 1
pcors <- ifelse(abs(pcors) < 0.05, 0, pcors)

inv <-  -pcors
diag(inv) <- 1
cors <- cov2cor( solve(inv))

# example data
Y <- BGGM::gen_ordinal(n = 500, p = 5,
                   levels = 2,
                    cor_mat = cors,
                    empirical = FALSE)
}
