% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm_compare_bf.default.R
\name{ggm_compare_bf.default}
\alias{ggm_compare_bf.default}
\title{Compare GGMs with Bayesian Hypothesis Testing}
\usage{
\method{ggm_compare_bf}{default}(..., prior_sd, iter = 25000,
  cores = 2)
}
\arguments{
\item{...}{data matrices (\emph{n} \times  \emph{p}). Requires at least two.}

\item{prior_sd}{hypothesized standard deviation for the edges or partial correlations.}

\item{iter}{number of posterior samples.}

\item{cores}{number of cores for parallel computing. The default is 2, but this can be adjusted.}
}
\value{
an object of class \code{ggm_compare_bf}
}
\description{
Comparing any number of GGMs with Bayesian hypothesis testing (i.e., with the Bayes factor). This method allows for
assessing (relative) evidence for edge equality or edges differences across any number of groups.
}
\examples{

# assume null is true
Y1 <- MASS::mvrnorm(500, rep(0, 16), Sigma = diag(16))
Y2 <- MASS::mvrnorm(500, rep(0, 16), Sigma = diag(16))
Y3 <- MASS::mvrnorm(500, rep(0, 16), Sigma = diag(16))

bf_ggm <- ggm_compare_bf(Y1, Y2, Y3, prior_sd = .5, iter = 2500, cores = 2)
select( bf_ggm, BF_cut = 3)
}
