% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGM_compare_bf.R
\name{GGM_compare_bf}
\alias{GGM_compare_bf}
\title{GGM_compare_bf}
\usage{
GGM_compare_bf(X1, X2, hyp, priorprob = 1, cores, delta, nu, n_samples)
}
\arguments{
\item{n_samples}{}
}
\description{
GGM_compare_bf
}
