% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm_compare_confirm.R
\name{ggm_compare_confirm}
\alias{ggm_compare_confirm}
\title{GGM Compare: Confirmatory Hypothesis Testing}
\usage{
ggm_compare_confirm(
  ...,
  hypothesis,
  formula = NULL,
  type = "continuous",
  mixed_type = NULL,
  prior_sd = 0.2,
  iter = 5000
)
}
\arguments{
\item{...}{at least two matrices (or data frame) of dimensions \emph{n} (observations) by  \emph{p} (nodes).}

\item{hypothesis}{character string. hypothesis (or hypotheses) to be tested. See notes for futher details.}

\item{formula}{an object of class \code{\link[stats]{formula}}. This allows for including
control variables in the model (i.e., \code{~ gender}).}

\item{type}{character string. Which type of data for \strong{Y} ? The options include \code{continuous},
\code{binary}, or \code{ordinal}. See the note for further details.}

\item{mixed_type}{numeric vector. An indicator of length p for which varibles should be treated as ranks.
(1 for rank and 0 to assume normality). The default is currently (dev version) to treat all integer variables
as ranks when \code{type = "mixed"} and \code{NULL} otherwise. See note for further details.}

\item{prior_sd}{The scale of the prior distribution (centered at zero), in reference to a beta distribtuion.
The `default` is 0.25. See note for further details.}

\item{iter}{number of iterations (posterior samples; defaults to 5000).}

\item{seed}{The seed for random number generation (default set to \code{1}).}
}
\value{

}
\description{
GGM Compare: Confirmatory Hypothesis Testing
}
