% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm_compare_estimate.default.R
\name{ggm_compare_estimate}
\alias{ggm_compare_estimate}
\title{GGM Compare: Estimate}
\usage{
ggm_compare_estimate(
  ...,
  formula = NULL,
  data = NULL,
  type = "continuous",
  mixed_type = NULL,
  analytic = FALSE,
  prior_sd = 0.5,
  iter = 5000
)
}
\arguments{
\item{...}{matrices (or data frame) of dimensions \emph{n} (observations) by  \emph{p} (variables).
Requires at least two.}

\item{formula}{an object of class \code{\link[stats]{formula}}. This allows for including
control variables in the model (i.e., \code{~ gender}).}

\item{data}{optional data frames in a list that containing the variables in \code{formula}.
This is required when controlling for variables. Requires at least two (one for each group).
See \code{examples} for futher details.}

\item{type}{character string. Which type of data for \strong{Y} ? The options include \code{continuous},
\code{binary}, or \code{ordinal}. See the note for further details.}

\item{mixed_type}{numeric vector. An indicator of length p for which varibles should be treated as ranks.
(1 for rank and 0 to assume normality). The default is currently (dev version) to treat all integer variables
as ranks when \code{type = "mixed"} and \code{NULL} otherwise. See note for further details.}

\item{analytic}{logical. Should the analytic solution be computed (default is \code{FALSE}) ?}

\item{prior_sd}{The scale of the prior distribution (centered at zero), in reference to a beta distribtuion.
The `default` is 0.50. See note for further details.}

\item{iter}{number of iterations (posterior samples; defaults to 5000).}
}
\value{
A list of class \code{ggm_compare_estimate} containing:
 \itemize{
 \item \code{pcor_diffs} partial correlation differences (posterior distribution)
 \item \code{p} number of variable
 \item \code{info} list containing information about each group (e.g., sample size, etc.)
 \item \code{iter} number of posterior samples
 \item \code{call} \code{match.call}
 }
}
\description{
Compare partial correlations that are estimated from any number of groups. This method works for
continuous, binary, ordinal, and mixed data.  The approach (i.e., a difference between posterior distributions) was
described in  \insertCite{Williams2019;textual}{BGGM}.
}
\note{
This function can be used to compare the partial correlations for any number of groups.
This is accomplished with pairwise comparisons for each relation. In the case of three groups,
for example, group 1 and group 2 are compared, then group 1 and group 3 are compared, and then
group 2 and group 3 are compared. There is a full distibution for each difference that can be
summarized  \code{\link{summary.ggm_compare_estimate}} and then visualized.
\code{\link{plot.summary.ggm_compare_estimate}}
}
\examples{
# comparing two groups (males vs. females) in a personality network




}
\references{
\insertAllCited{}
}
