% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm_compare_ppc.default.R
\name{ggm_compare_ppc.default}
\alias{ggm_compare_ppc.default}
\title{Compare GGMs with Posterior Predictive KL-Divergence}
\usage{
\method{ggm_compare_ppc}{default}(..., type = "global", iter = 5000,
  cores = 2)
}
\arguments{
\item{...}{data matrices (\emph{n} \times  \emph{p}). Requires at least two.}

\item{type}{\code{type = "global"} for testing the entire precision matrix. \code{type = "nodewise"} for testing each node (i.e., variable)}

\item{iter}{number of replicated data sets. default is \code{iter = 1000}.}

\item{cores}{Number of cores for parallel computing. The defaul is 2, but this can be adjusted.}
}
\value{
object of class \code{ggm_compare_ppc}
}
\description{
Compare GGMs with Posterior Predictive KL-Divergence
}
\examples{

# Assume null is true
Y1 <- MASS::mvrnorm(500, rep(0, 16), Sigma = BGGM::ptsd_cor1)
Y2 <- MASS::mvrnorm(500, rep(0, 16), Sigma = BGGM::ptsd_cor1)


ggm_ppc <- ggm_compare_ppc(Y1, Y2, iter = 1000)
}
