% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm_compare_ppc.default.R
\name{ggm_compare_ppc}
\alias{ggm_compare_ppc}
\title{S3 estimate method}
\usage{
ggm_compare_ppc(...)
}
\arguments{
\item{...}{currently not used}
}
\description{
S3 estimate method
}
\seealso{
\code{\link{ggm_compare_ppc.default}}
}
