% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gss}
\alias{gss}
\title{Data: 1994 General Social Survey}
\format{
A data frame containing 1190 observations (n = 1190) and 6 variables (p = 6) measured on the binary scale
        \insertCite{fowlkes1988evaluating}{BGGM}. The variable descriptions were copied from
        \insertCite{@section 4, @hoff2007extending;textual}{BGGM}
}
\usage{
data("gss")
}
\description{
A data frame containing 1002 rows and 7 variables measured on various scales,
including binary and ordered cateogrical (with varying numbers of categories).
There are also missing values in each variable

\itemize{
  \item \code{Inc}  Income of the respondent in 1000s of dollars, binned into 21 ordered categories.
  \item \code{DEG}   Highest degree ever obtained (none, HS, Associates, Bachelors, or Graduate)
  \item \code{CHILD}  Number of children ever had.
  \item \code{PINC}  Financial status of respondent's parents when respondent was 16 (on a 5-point scale).
  \item \code{PDEG}  Maximum of mother's and father's highest degree
  \item \code{PCHILD}  Number of siblings of the respondent plus one
  \item \code{AGE} Age of the respondent in years.
}
}
\examples{
data("gss")
}
\references{
\insertAllCited{}
}
\keyword{datasets}
