% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_plot.R
\name{hypothesis_plot}
\alias{hypothesis_plot}
\title{Hypothesis Plot}
\usage{
hypothesis_plot(rho_sd = NULL, difference = FALSE, fit = NULL,
  edge = NULL, samples = 10000, size = NULL)
}
\arguments{
\item{size}{}
}
\description{
Hypothesis Plot
}
