% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{mae}
\alias{mae}
\title{Mean Absolute Error}
\usage{
mae(object, ...)
}
\arguments{
\item{object}{object of class \code{post.pred} or  \code{predict.estimate}}

\item{...}{currently ignored}
}
\value{
object of class \code{metric}
}
\description{
Mean Absolute Error
}
\examples{
# data
Y <- subset(tas, gender == "M")[,-ncol(tas)]

# fit model
fit <- estimate(Y)

# predict (note summary = FALSE)
pred <- predict(fit, iter = 50, summary = FALSE)

mae(pred)
}
