% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\title{Maximum A Posteriori Precision Matrix}
\usage{
map(Y)
}
\arguments{
\item{Y}{data matrix (\emph{n} by  \emph{p})}
}
\value{
object of class \code{map}
}
\description{
Maximum A Posteriori Precision Matrix
}
\examples{
# p = 20
Y <- BGGM::bfi[, 1:5]

# map
map <- map(Y)
map
}
