% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{mse}
\alias{mse}
\title{Mean Square Error for \code{predict.estimate} Objects}
\usage{
mse(object, ...)
}
\arguments{
\item{object}{object of class \code{predict.estimate}}

\item{...}{currently ignored}
}
\value{

}
\description{
Mean Square Error for \code{predict.estimate} Objects
}
\examples{
# data
Y <- subset(tas, gender == "M")[,-ncol(tas)]

# fit model
fit <- estimate(Y)

# predict (note summary = FALSE)
pred <- predict(fit, summary = FALSE)

mse(pred)
}
