% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_measures.R
\name{performance_measures}
\alias{performance_measures}
\title{Title}
\usage{
performance_measures(ThetaEst, Theta)
}
\arguments{
\item{ThetaEst}{Estimated adjacency matrix}

\item{Theta}{True adjacency matrix}
}
\description{
Title
}
