% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictability.R
\name{plot.assess_predictability}
\alias{plot.assess_predictability}
\title{Plot \code{assess_predictability} Objects}
\usage{
\method{plot}{assess_predictability}(x, ...)
}
\arguments{
\item{x}{object of class \code{assess_predictability}}
}
\value{

}
\description{
Plot \code{assess_predictability} Objects
}
