% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.edge_compare.estimate.R
\name{plot.compare.estimate}
\alias{plot.compare.estimate}
\title{Title}
\usage{
\method{plot}{compare.estimate}(x, size = 2, color = "red",
  width = 0.01, prob = NULL)
}
\arguments{
\item{prob}{}
}
\description{
Title
}
