% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.compare.predict.R
\name{plot.compare.predict}
\alias{plot.compare.predict}
\title{Plotting (node-wise) Differences in Predictive Accuracy}
\usage{
\method{plot}{compare.predict}(x, limits)
}
\arguments{
\item{x}{object of type \code{predict}}

\item{limits}{upper and lower boundary for gradient}
}
\description{
Plotting (node-wise) Differences in Predictive Accuracy
}
