% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.edge_compare.estimate.R
\name{plot.edge_compare.estimate}
\alias{plot.edge_compare.estimate}
\title{Plot Edge Contrasts from the Estimation Based Methods}
\usage{
\method{plot}{edge_compare.estimate}(x, size = 2, color = "red",
  width = 0.01, prob = NULL)
}
\arguments{
\item{prob}{}
}
\description{
Plot Edge Contrasts from the Estimation Based Methods
}
