% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.edge_compare.explore.R
\name{plot.edge_compare.explore}
\alias{plot.edge_compare.explore}
\title{Plot Edge Contrasts from the Bayes Factor Based Methods}
\usage{
\method{plot}{edge_compare.explore}(x, stack = TRUE, hjust1 = -1,
  hjust2 = -1, hjust3 = -1, base_text = 14, leg_text = 10,
  spread = 0.95)
}
\arguments{
\item{leg_text}{}
}
\description{
Plot Edge Contrasts from the Bayes Factor Based Methods
}
