% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.estimate.R
\name{plot.estimate}
\alias{plot.estimate}
\title{Plot Edges (Parial Correlation)}
\usage{
\method{plot}{estimate}(x, ci_width, size, width)
}
\arguments{
\item{width}{}
}
\description{
Plot Edges (Parial Correlation)
}
