% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.loocv.R
\name{plot.loocv}
\alias{plot.loocv}
\title{Plot Nodewise Leave-One-Out Cross-Validation}
\usage{
\method{plot}{loocv}(x, size = 2, color = "blue", width = 0.01)
}
\arguments{
\item{width}{}
}
\description{
Plot Nodewise Leave-One-Out Cross-Validation
}
