% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_imputation.R
\name{plot.mvn_imputation}
\alias{plot.mvn_imputation}
\title{Plot \code{mvn_imputation} Objects}
\usage{
\method{plot}{mvn_imputation}(x, index = NULL, cred = 0.95, alpha = 0.5, scale = 1, color = "blue", ...)
}
\arguments{
\item{x}{An object of class \code{mvn_imputation}}

\item{index}{Character string. Which missing value should be plotted (defaults to the first). Note
that the format is in referece to the row and column (e.g., \code{10--1}).
is the 10th row in the 1st column.}

\item{cred}{Numeric. The credible interval width for summarizing the posterior
distributions (defaults to 0.95; must be between 0 and 1).}

\item{alpha}{Numeric. Transparancey of the ridges (defaults to 0.50).}

\item{scale}{Numeric. This controls the overlap of densities (defaults to 1).}

\item{color}{Character string. What color for tail region (\code{type = "ridgeline"} ) ?
Defaults to \code{"blue"}.}

\item{...}{Currently ignored.}
}
\value{
An object of class \code{ggplot}.
}
\description{
Plot \code{mvn_imputation} Objects
}
\examples{
\donttest{
# obs
n <- 5000

# n missing
n_missing <- 1000

# variables
p <- 16

# data
Y <- MASS::mvrnorm(n, rep(0, p), ptsd_cor1)

# for checking
Ymain <- Y

# all possible indices
indices <- which(matrix(0, n, p) == 0,
                 arr.ind = TRUE)

# random sample of 1000 missing values
na_indices <- indices[sample(5:nrow(indices),
                             size = n_missing,
                             replace = FALSE),]

# fill with NA
Y[na_indices] <- NA

# missing = 1
Y_miss <- ifelse(is.na(Y), 1, 0)

# true values (to check)
true <- unlist(sapply(1:p, function(x)
        Ymain[which(Y_miss[,x] == 1),x] ))

# impute
fit_missing <- mvn_imputation(Y, progress = FALSE, iter = 250)

index <- paste0(na_indices[1:10,1], "--", na_indices[1:10,2])

plot(fit_missing, index = index)
}
}
