% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_sum.R
\name{plot.pcor_sum}
\alias{plot.pcor_sum}
\title{Plot \code{posterior_sum} Object}
\usage{
\method{plot}{pcor_sum}(x, fill = "#CC79A7", ...)
}
\arguments{
\item{x}{An object of class \code{posterior_sum}}

\item{fill}{Character string. What fill for the histogram
(defaults to colorblind "pink")?}
}
\value{
A list of \code{ggplot} objects
}
\description{
Plot \code{posterior_sum} Object
}
\note{
\strong{Examples}:
}
\seealso{
posterior_sum
}
