% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_check.R
\name{plot.robust_check}
\alias{plot.robust_check}
\title{Plot \code{robust_check} Objects}
\usage{
\method{plot}{robust_check}(x, size = 2, ...)
}
\arguments{
\item{x}{object of class \code{robust_check}}

\item{size}{geom point size}

\item{...}{currently ignored}
}
\value{
\code{ggplot} object
}
\description{
Plot \code{robust_check} Objects
}
\examples{
# all equal
Y1 <- MASS::mvrnorm(500, rep(0, 16), Sigma = diag(16))
Y2 <- MASS::mvrnorm(500, rep(0, 16), Sigma = diag(16))
Y3 <- MASS::mvrnorm(500, rep(0, 16), Sigma = diag(16))

# two priors
prior_10 <- ggm_compare_bf(Y1, Y2, Y3, prior_sd = .10, iter = 250, cores = 2)
prior_20 <- ggm_compare_bf(Y1, Y2, Y3, prior_sd = .20, iter = 250, cores = 2)

# robustness check
rc <- robust_check(prior_10, prior_20)

# print results
rc
# plot results
plot(rc)
}
