% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.select.ggm_compare_bf.R
\name{plot.select.ggm_compare_bf}
\alias{plot.select.ggm_compare_bf}
\title{Plot Edge Differences and Equivalence Between GGMs (\code{ggm_compare_bf })}
\usage{
\method{plot}{select.ggm_compare_bf}(x, mat_type = "adj",
  type = "network", lower_tri = TRUE, layout = "circle",
  node_outer = 10, node_inner = 9, node_text_size = 8,
  labels = NULL)
}
\arguments{
\item{x}{object of class \code{select.ggm_compare_bf}}

\item{mat_type}{"adj" or "BF"}

\item{type}{\code{heatmap} or \code{network} plot (network is currently the only option)}

\item{lower_tri}{heatmap lower triangular elements}

\item{layout}{layout of network plot. see igraph for options}

\item{node_outer}{border for node}

\item{node_inner}{node size (see notes)}

\item{node_text_size}{node text size}

\item{labels}{node labels. Default is 1:p, but can change to colnames(Y)}
}
\description{
Plot Edge Differences and Equivalence Between GGMs (\code{ggm_compare_bf })
}
