% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.select.ggm_compare_estimate.R
\name{plot.select.ggm_compare_estimate}
\alias{plot.select.ggm_compare_estimate}
\title{Plot Edge Differences and Equivalence Between GGMs (\code{ggm_compare_estimate})}
\usage{
\method{plot}{select.ggm_compare_estimate}(x, type = "network",
  lower_tri = TRUE, layout = "circle", node_outer = 10,
  node_inner = 9, node_text_size = 8, labels = NULL)
}
\arguments{
\item{x}{object of class \code{select.ggm_compare_estimate}}

\item{type}{\code{heatmap} or \code{network} plot (network is currently the only option)}

\item{lower_tri}{heatmap lower triangular elements}

\item{layout}{layout of network plot. see igraph for options}

\item{node_outer}{border for node}

\item{node_inner}{node size (see notes)}

\item{node_text_size}{node text size}

\item{labels}{node labels. Default is 1:p, but can change to colnames(Y)}
}
\description{
Plot Edge Differences and Equivalence Between GGMs (\code{ggm_compare_estimate})
}
\examples{
Y1 <- MASS::mvrnorm(5000, rep(0, 16), Sigma = diag(16))
Y2 <- MASS::mvrnorm(5000, rep(0, 16), Sigma = diag(16))
Y3 <- MASS::mvrnorm(5000, rep(0, 16), Sigma = diag(16))
ggm_est <- ggm_compare_estimate(Y1, Y2, Y3,  iter = 5000, ci_width = .95)

sel <- select(ggm_est, type = "ci")

plot(x, labels = letters[1:16])
}
