% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{plot.summary.estimate}
\alias{plot.summary.estimate}
\title{Plot \code{summary.estimate}}
\usage{
\method{plot}{summary.estimate}(x, ...)
}
\arguments{
\item{x}{an object of class \code{summary.estimate}}

\item{...}{currently ignored}
}
\value{
an object of class \code{ggplot}
}
\description{
Plot \code{summary.estimate}
}
