% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.explore.R
\name{plot.summary.select.explore}
\alias{plot.summary.select.explore}
\title{GGM: Plot \code{summary.select.explore] Objects}}
\usage{
\method{plot}{summary.select.explore}(x, size = 2, color = "black")
}
\arguments{
\item{x}{object of class \code{summary.select.explore}}

\item{size}{numeric. Size of the points.}

\item{color}{string. Color of the points.}
}
\value{

}
\description{
Visualize posterior hypothesis probabilities.
}
