% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_predict.R
\name{posterior_predict}
\alias{posterior_predict}
\title{Posterior Predictive Distribution for \code{estimate} Objects}
\usage{
posterior_predict(
  object,
  iter = 500,
  cred = 0.95,
  newdata = NULL,
  summary = TRUE,
  ...
)
}
\arguments{
\item{object}{object of class \code{estimate}}

\item{iter}{iters used to compute the residuals}

\item{cred}{credible interval used for summarizing}

\item{newdata}{an optional data frame for obtaining predictions (e.g., on test data)}

\item{summary}{summarize the posterior samples (Default is \code{TRUE}).
Setting it to \code{FALSE} can be used to then compute performance metrics.}

\item{...}{}
}
\value{

}
\description{
Posterior Predictive Distribution for \code{estimate} Objects
}
