% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_samples.R
\name{posterior_samples}
\alias{posterior_samples}
\title{Extract Posterior Samples}
\usage{
posterior_samples(object, ...)
}
\arguments{
\item{object}{an object of class \code{estimate} or \code{explore}.}

\item{...}{currently ignored.}
}
\value{
A matrix of posterior samples for the partial correlation. Note that if controlling for
        variables (e.g., formula \code{~ age}), the matrix also includes the coefficients from each
        multivariate regression.
}
\description{
Extract posterior samples for all parameters.
}
