% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision.R
\name{precision}
\alias{precision}
\title{Precision Matrix Posterior Distribution}
\usage{
precision(object, progress = TRUE)
}
\arguments{
\item{object}{An object of class \code{estimate}.}

\item{progress}{Logical. Should a progress bar be included (defaults to \code{TRUE}) ?}
}
\value{
A 3d array of dimensions \code{p} by \code{p} by \code{iter} including
**unconstrained** (i.e., from th full graph) precision matrices.
}
\description{
Transform the sampled correlation matrices to
precision matrices (i.e., inverse covariance matrices).
}
\examples{
\donttest{
# data
Y <- ptsd

# fit model
fit <- estimate(Y)

# precisio matrix
Theta <- precision(fit)

}

}
