% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.estimate.R
\name{predict.explore}
\alias{predict.explore}
\title{Model Predictions for \code{explore} Objects}
\usage{
\method{predict}{explore}(object, newdata = NULL, summary = TRUE, cred = 0.95, iter = NULL, ...)
}
\arguments{
\item{object}{object of class \code{explore}}

\item{newdata}{an optional data frame for obtaining predictions (e.g., on test data)}

\item{summary}{summarize the posterior samples (defaults to \code{TRUE}).}

\item{cred}{credible interval used for summarizing}

\item{iter}{number of posterior samples (defaults to all in the object).}

\item{...}{currently ignored}
}
\value{
\code{summary = TRUE}: 3D array of dimensions n (observations),
        4 (posterior summary),
        p (number of nodes). \code{summary = FALSE}:
        list containing predictions for each variable
}
\description{
Model Predictions for \code{explore} Objects
}
\examples{
\donttest{
# data
Y <- subset(tas, gender == "M")[,-ncol(tas)]

# fit model
fit <- estimate(Y)

# predict
predict(fit, iter = 25)
}
}
