% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictability.R
\name{predictability}
\alias{predictability}
\title{Nodewise Predicatability}
\usage{
predictability(object, cred = 0.95, iter = 1000, cores = 2, ...)
}
\arguments{
\item{object}{object of class \code{estimate}}

\item{cred}{credible interval width used for the decision rule}

\item{iter}{iterations used for computing R2}

\item{cores}{number of cores for parallel computing}

\item{...}{currently ignored}
}
\value{

}
\description{
Nodewise Predicatability
}
