% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_error.R
\name{prediction_error}
\alias{prediction_error}
\title{Title}
\usage{
prediction_error(fit, measure = c("KL"), newdata = NULL, selected,
  ci_width = 0.95, samples = 500)
}
\arguments{
\item{samples}{}
}
\description{
Title
}
