% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_BGGM.R
\name{print.BGGM}
\alias{print.BGGM}
\title{Print method for \code{BGGM} objects}
\usage{
\method{print}{BGGM}(x, ...)
}
\arguments{
\item{x}{An object of class \code{BGGM}}

\item{...}{currently ignored}
}
\description{
Print method for \code{BGGM} objects
}
