% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.estimate.R
\name{print.coef.estimate}
\alias{print.coef.estimate}
\title{Print method for \code{coef.estimate} objects}
\usage{
\method{print}{coef.estimate}(x, ...)
}
\arguments{
\item{x}{An object of class \code{coef.estimate}}

\item{...}{currently ignored}
}
\description{
Print method for \code{coef.estimate} objects
}
