% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirm.R
\name{print.confirm}
\alias{print.confirm}
\title{Print Method for \code{confirm} Objects}
\usage{
\method{print}{confirm}(x, ...)
}
\arguments{
\item{x}{object of class \code{confirm}}

\item{...}{currently ignored}
}
\description{
Print Method for \code{confirm} Objects
}
