% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{print.estimate}
\alias{print.estimate}
\title{Print method for \code{estimate.default} objects}
\usage{
\method{print}{estimate}(x, ...)
}
\arguments{
\item{x}{An object of class \code{estimate}}

\item{...}{currently ignored}
}
\description{
Print method for \code{estimate.default} objects
}
\examples{
# data
Y <- BGGM::bfi[, 1:5]
# analytic approach (sample by setting analytic = FALSE)
fit <- estimate(Y, analytic = TRUE)
fit
}
\seealso{
\code{\link{select.estimate}}
}
