% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.default.R
\name{print.explore}
\alias{print.explore}
\title{Print method for \code{explore.default} objects}
\usage{
\method{print}{explore}(x, ...)
}
\arguments{
\item{x}{An object of class \code{explore}}

\item{...}{currently ignored}
}
\description{
Print method for \code{explore.default} objects
}
