% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{print.fitted.estimate}
\alias{print.fitted.estimate}
\title{Print Method for \code{fitted.estimate} Objects}
\usage{
\method{print}{fitted.estimate}(x, ...)
}
\arguments{
\item{x}{object of class \code{fitted.estimate}}

\item{...}{currently ignored}
}
\description{
Print Method for \code{fitted.estimate} Objects
}
