% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm_compare_bf.default.R
\name{print.ggm_compare_bf}
\alias{print.ggm_compare_bf}
\title{Print method for \code{ggm_compare_bf} objects}
\usage{
\method{print}{ggm_compare_bf}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ggm_compare_bf}}

\item{...}{currently ignored}
}
\description{
Print method for \code{ggm_compare_bf} objects
}
\seealso{
\code{\link{ggm_compare_bf.default}}
}
