% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm_compare_estimate.default.R
\name{print.ggm_compare_estimate}
\alias{print.ggm_compare_estimate}
\title{Print method for \code{ggm_compare_estimate} objects}
\usage{
\method{print}{ggm_compare_estimate}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ggm_compare_estimate}}

\item{...}{currently ignored}
}
\description{
Print method for \code{ggm_compare_estimate} objects
}
\seealso{
\code{\link{ggm_compare_estimate.default}}
}
