% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm_compare_ppc.default.R
\name{print.ggm_compare_ppc}
\alias{print.ggm_compare_ppc}
\title{Print method for \code{ggm_compare_ppc.default} objects}
\usage{
\method{print}{ggm_compare_ppc}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ggm_compare_ppc}}

\item{...}{currently ignored}
}
\description{
Print method for \code{ggm_compare_ppc.default} objects
}
