% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{print.map}
\alias{print.map}
\title{Print Method for \code{map} Objects}
\usage{
\method{print}{map}(x, ...)
}
\arguments{
\item{x}{object of class map}

\item{...}{currently ignored}
}
\description{
Print Method for \code{map} Objects
}
