% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{print.metric}
\alias{print.metric}
\title{Print Method for \code{metric} Objects}
\usage{
\method{print}{metric}(x, ...)
}
\arguments{
\item{x}{object of class \code{metric}}

\item{...}{currently ignored}
}
\value{

}
\description{
Print Method for \code{metric} Objects
}
