% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_predict.R
\name{print.post.pred}
\alias{print.post.pred}
\title{Print Method for \code{post.pred} Objects}
\usage{
\method{print}{post.pred}(x, ...)
}
\arguments{
\item{x}{object of class \code{post.pred}}

\item{...}{currently ignored}
}
\description{
Print Method for \code{post.pred} Objects
}
