% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.estimate.R
\name{print.predict.estimate}
\alias{print.predict.estimate}
\title{Print Method for \code{predict.estimate} Objects}
\usage{
\method{print}{predict.estimate}(x, ...)
}
\arguments{
\item{x}{object of class \code{predict.estimate}}

\item{...}{currently ignored}
}
\description{
Print Method for \code{predict.estimate} Objects
}
