% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.estimate.R
\name{print.predict}
\alias{print.predict}
\title{Print Method for \code{predict} Objects}
\usage{
\method{print}{predict}(x, ...)
}
\arguments{
\item{x}{object of class \code{predict}}

\item{...}{currently ignored}
}
\description{
Print Method for \code{predict} Objects
}
