% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_check.R
\name{print.robust_check}
\alias{print.robust_check}
\title{Print method for \code{robust_check} objects}
\usage{
\method{print}{robust_check}(x, ...)
}
\arguments{
\item{x}{object of class \code{robust_check}}

\item{...}{currently ignored}
}
\description{
Print method for \code{robust_check} objects
}
