% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.estimate.R
\name{print.select.estimate}
\alias{print.select.estimate}
\title{Print method for \code{select.estimate} objects}
\usage{
\method{print}{select.estimate}(x, ...)
}
\arguments{
\item{x}{An object of class \code{select.estimate}}

\item{...}{currently ignored}
}
\description{
Print method for \code{select.estimate} objects
}
\seealso{
\code{\link{select.estimate}}
}
