% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.explore.R
\name{print.select.explore}
\alias{print.select.explore}
\title{Print method for \code{select.explore} objects}
\usage{
\method{print}{select.explore}(x, summarize = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{select.explore}}

\item{summarize}{summarize connectivity}

\item{...}{currently ignored}
}
\description{
Print method for \code{select.explore} objects
}
\seealso{
\code{\link{select.explore}}
}
