% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.ggm_compare_estimate.R
\name{print.select.ggm_compare_estimate}
\alias{print.select.ggm_compare_estimate}
\title{Print method for \code{select.ggm_compare_estimate} objects}
\usage{
\method{print}{select.ggm_compare_estimate}(x, ...)
}
\arguments{
\item{x}{An object of class \code{select.ggm_compare_estimate}}

\item{...}{currently ignored}
}
\description{
Print method for \code{select.ggm_compare_estimate} objects
}
\seealso{
\code{\link{select.ggm_compare_estimate}}
}
