% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{print.summary.estimate}
\alias{print.summary.estimate}
\title{Summary method for \code{summary.estimate} objects}
\usage{
\method{print}{summary.estimate}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.estimate}}

\item{...}{currently ignored}
}
\description{
Summary method for \code{summary.estimate} objects
}
\seealso{
\code{\link{summary.estimate}}
}
