% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.default.R
\name{print.summary.explore}
\alias{print.summary.explore}
\title{Summary method for \code{summary.explore} objects}
\usage{
\method{print}{summary.explore}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.explore}}

\item{...}{currently ignored}
}
\description{
Summary method for \code{summary.explore} objects
}
\seealso{
\code{\link{summary.explore}}
}
