% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm_compare_ppc.default.R
\name{print.summary.ggm_compare_ppc}
\alias{print.summary.ggm_compare_ppc}
\title{Summary method for \code{summary.ggm_compare_ppc} objects}
\usage{
\method{print}{summary.ggm_compare_ppc}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.ggm_compare_ppc}}

\item{...}{currently ignored}
}
\description{
Summary method for \code{summary.ggm_compare_ppc} objects
}
\seealso{
\code{\link{summary.ggm_compare_ppc}}
}
