% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm_compare_bf.default.R
\name{print.summary.ggm_estimate_bf}
\alias{print.summary.ggm_estimate_bf}
\alias{print.summary.ggm_compare_bf}
\title{Summary method for \code{summary.ggm_compare_bf} objects}
\usage{
\method{print}{summary.ggm_compare_bf}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.ggm_compare_bf}}

\item{...}{currently ignored}
}
\description{
Summary method for \code{summary.ggm_compare_bf} objects
}
\seealso{
\code{\link{ggm_compare_bf}}
}
