% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm_compare_estimate.default.R
\name{print.summary.ggm_estimate_compare}
\alias{print.summary.ggm_estimate_compare}
\title{Summary method for \code{summary.ggm_compare_estimate} objects}
\usage{
\method{print}{summary.ggm_estimate_compare}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.ggm_compare_estimate}}

\item{...}{currently ignored}
}
\description{
Summary method for \code{summary.ggm_compare_estimate} objects
}
\examples{
# data
Y1 <- BGGM::bfi[1:500,1:5]
Y2 <- BGGM::bfi[501:1000, 1:5]

# fit model
fit <- ggm_compare_estimate(Y1, Y2)

# posterior summary of differences
summary(fit)
}
\seealso{
\code{\link{ggm_compare_estimate.default}}
}
