% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{print.summary.metric}
\alias{print.summary.metric}
\title{Print Method for \code{summary.metric} Object}
\usage{
\method{print}{summary.metric}(x, digits = 2, ...)
}
\arguments{
\item{x}{object of class \code{summary.metric}}

\item{digits}{digits used to round the values}

\item{...}{currently ignored}
}
\description{
Print Method for \code{summary.metric} Object
}
