% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ptsd}
\alias{ptsd}
\title{Data: Post-Traumatic Stress Disorder}
\format{
A dataframe with 221 rows and 20 variables
}
\usage{
data("ptsd")
}
\description{
A dataset containing items that measure Post-traumatic stress disorder symptoms \insertCite{armour2017network}{BGGM}.
There are 20 variables (\emph{p}) and  221 observations (\emph{n}).
}
\details{
\itemize{
  \item Intrusive Thoughts
  \item Nightmares
  \item Flashbacks
  \item Emotional cue reactivity
  \item Psychological cue reactivity
  \item Avoidance of thoughts
  \item Avoidance of reminders
  \item Trauma-related amnesia
  \item Negative beliefs
  \item Negative trauma-related emotions
  \item Loss of interest
  \item Detachment
  \item Restricted affect
  \item Irritability/anger
  \item Self-destructive/reckless behavior
  \item Hypervigilance
  \item Exaggerated startle response
  \item Difficulty concentrating
  \item Sleep disturbance
}
}
\references{
\insertAllCited{}
}
\keyword{datasets}
