% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_check.R
\name{robust_check}
\alias{robust_check}
\title{Robustness Check for \code{ggm_compare_bf} Objects}
\usage{
robust_check(..., BF_cut = 3)
}
\arguments{
\item{...}{objects of class \code{ggm_compare_bf} (at least two)}

\item{BF_cut}{evidentiary threshold (default is 3)}
}
\value{
object of class \code{robust_check}
}
\description{
Robustness Check for \code{ggm_compare_bf} Objects
}
\examples{
# all equal
Y1 <- MASS::mvrnorm(500, rep(0, 16), Sigma = diag(16))
Y2 <- MASS::mvrnorm(500, rep(0, 16), Sigma = diag(16))
Y3 <- MASS::mvrnorm(500, rep(0, 16), Sigma = diag(16))

# two priors
prior_10 <- ggm_compare_bf(Y1, Y2, Y3, prior_sd = .10, iter = 250, cores = 2)
prior_20 <- ggm_compare_bf(Y1, Y2, Y3, prior_sd = .20, iter = 250, cores = 2)

# robustness check
rc <- robust_check(prior_10, prior_20)

# print results
rc
# plot results
plot(rc)
}
