% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_your_own.R
\name{roll_your_own}
\alias{roll_your_own}
\title{Compute Custom Network Statistics}
\usage{
roll_your_own(object, FUN, select = FALSE, cred = 0.95, extract = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{estimate}.}

\item{FUN}{A custom function for computing the statistic.}

\item{select}{Logical. Should the graph be selected ? The default is currently \code{FALSE}.}

\item{cred}{Numeric. Credible interval between 0 and 1  (default is 0.95) that is used for selecting the graph.}

\item{extract}{A character specifying what, if anything, should be extracted from the function. For example,
to extract the object "r", set \code{extract = "r"} and internally \code{$r} will be added
to the end of the function.}

\item{...}{Currently ignored}
}
\value{
An object defined by \code{FUN}.
}
\description{
This function allows for computing custom network statistics for
weighted adjacency matrices (partial correlations). The statistics are computed for
each of the sampled matrices, resulting in a distribution.
}
\details{
The user has complete control of this function. Hence, care must be taken as to what \code{FUN} returns
and in what format.
}
\examples{
\donttest{

####################################
###### example 1: assortment #######
####################################
library(assortnet)

Y <- BGGM::bfi[,1:10]
membership <- c(rep("a", 5), rep("c", 5))

# fit model
fit <- estimate(Y = Y,
                analytic = FALSE,
                iter = 1000)

# list of columns belowinging in each group
e.g., first 5 are "a", last 5 are "c"

membership <- c(rep("a", 5), rep("c", 5))


net_stat <- roll_your_own(object = fit,
                          FUN = assortment.discrete,
                          types = membership, weighted = TRUE,
                          SE = FALSE, M = 1,
                          extract = "r")

hist(net_stat)
}


}
