% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{rsa}
\alias{rsa}
\title{Resilience Scale of Adults (RSA)}
\format{A data frame with 28 variables and 1973 observations (5 point Likert scale)}
\usage{
data("rsa")
}
\description{
A dataset containing items from the Resilience Scale of Adults (RSA). There are 33 items  and
675 observations
}
\details{
\itemize{
  \item \code{1}  My plans for the future are
  \item \code{2}  When something unforeseen happens
  \item \code{3}  My family understanding of what is important in life is
  \item \code{4}  I feel that my future looks
  \item \code{5}  My goals
  \item \code{6}  I can discuss personal issues with
  \item \code{7}  I feel
  \item \code{8}  I enjoy being
  \item \code{9}  Those who are good at encouraging are
  \item \code{10} The bonds among my friends
  \item \code{11} My personal problems
  \item \code{12} When a family member experiences a crisis/emergency
  \item \code{13} My family is characterised by
  \item \code{14} To be flexible in social settings
  \item \code{15} I get support from
  \item \code{16} In difficult periods my family
  \item \code{17} My judgements and decisions
  \item \code{18} New friendships are something
  \item \code{19} When needed, I have
  \item \code{20} I am at my best when I
  \item \code{21} Meeting new people is
  \item \code{22} When I am with others
  \item \code{23} When I start on new things/projects
  \item \code{24} Facing other people, our family acts
  \item \code{25} Belief in myself
  \item \code{26} For me, thinking of good topics of conversation is
  \item \code{27} My close friends/family members
  \item \code{28} I am good at
  \item \code{29} In my family, we like to
  \item \code{30} Rules and regular routines
  \item \code{31} In difficult periods I have a tendency to
  \item \code{32} My goals for the future are
  \item \code{33} Events in my life that I cannot influence
  \item \code{gender} "M" (male) or "F" (female)

}
}
\note{
There are 6 domains

Planned future: items 1, 4, 5, 32

Perception of self: items 2, 11, 17, 25, 31, 33

Family cohesion: items 3, 7, 13, 16, 24, 29

Social resources: items 6, 9, 10, 12, 15, 19, 27

Social Competence: items 8, 14, 18, 21, 22, 26,

Structured style: items 23, 28, 30
}
\examples{
data("rsa")

# labels
rsa_labels <- BGGM:::rsa_labels
}
\references{
Briganti, G., & Linkowski, P. (2019). Item and domain network structures of the Resilience
Scale for Adults in 675 university students. Epidemiology and psychiatric sciences, 1-9.
}
\keyword{datasets}
