% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Sachs}
\alias{Sachs}
\title{Data: Sachs Network}
\format{
A data frame containing 7466 cells (n = 7466) and flow cytometry
 measurements of 11 (p = 11) phosphorylated proteins and phospholipids

 @references
 Sachs, K., Gifford, D., Jaakkola, T., Sorger, P., & Lauffenburger, D. A. (2002).
 Bayesian network approach to cell signaling pathway modeling. Sci. STKE, 2002(148), pe38-pe38.
}
\usage{
data("Sachs")
}
\description{
Protein expression in human immune system cells
}
\examples{
data("Sachs")

}
\keyword{datasets}
