% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.estimate.R
\name{select.estimate}
\alias{select.estimate}
\title{GGM Selection: Estimation}
\usage{
\method{select}{estimate}(object, cred = 0.95, alternative = "two.sided", ...)
}
\arguments{
\item{object}{object of class \code{estimate.default}}

\item{cred}{credible interval width used for the decision rule}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "two.sided" (default), "greater"  or "less".
See note for futher details.}

\item{...}{Not currently used}
}
\value{
An object of class \code{select.estimate}:

\itemize{
\item \code{pcor_adj} selected partial correlation matrix
\item \code{adj} adjacency matrix for the selected edges
\item \code{ci} credible interval width
\item \code{object} object of class \code{estimate}
}
}
\description{
Select the edge set for \code{estimate} objects. The graph is determined with
the posterior distribution, in particular credible interval exclusion of zero. This
corresponds to a directional posterior probability \insertCite{Marsman2017a}{BGGM}. For example,
the probability (conditional on the model) of a positive edge is at least 97.5% when the 95%
credible interval excludes zero.
}
\note{
\strong{Alternative Default:}

This package was built for the social-behavioral sciences in particular. In these applications, there is
strong theory that expects \emph{all} effects to be positive. This is known as a "positive manifold" and
this notion has a rich tradition in psychometrics. Hence, this can be incorportated into the graph with
\code{alternative = "greater"}. This results in the estimted structure including only positive edges. Further
details can be found at the blog "Dealing with Negative (Red) Edges in Psychological Networks: Frequentist Edition"
(\href{https://donaldrwilliams.github.io/2020/03/29/dealing-with-negative-red-edges-in-psychological-networks-frequentist-edition/}{link})
}
\examples{

# Analytic = TRUE
# p = 5
Y <- BGGM::bfi[,1:5]

# analytic solution
fit_analytic <- estimate(Y, analytic = TRUE)

# select E
E <- select(fit_analytic, ci_width = 0.95)

# non-zero partial correlations
E$partials_non_zero

# adjacency matrix
E$adjacency_non_zero
}
\references{
\insertAllCited{}
}
