% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.ggm_compare_estimate.R
\name{select.ggm_compare_estimate}
\alias{select.ggm_compare_estimate}
\title{Edge Differences and (Practical) Equivalence Between GGMs}
\usage{
\method{select}{ggm_compare_estimate}(x, type, prob = NULL)
}
\arguments{
\item{x}{object of class \code{ggm_compare_estimate}}

\item{type}{decision rule to use. either \code{"ci"} or \code{"rope"}}

\item{prob}{posterior probability. only used for \code{type = "rope"}}
}
\value{
list of class \code{select.ggm_compare_estimate}:

\code{type = "ci"}:
 \itemize{
 \item \code{mats_diff} adjacency matrices for edges differences (one for each group contrast)
 \item  \code{type} \code{ci} (used internally for summarizing and plotting)
 \item  \code{prob} \code{NULL} (used internally for summarizing and plotting)
 }
\code{type = "rope"}:
 \itemize{
 \item \code{mats_diff} adjacency matrices for edge differences (one for each group contrast)
 \item  \code{mats_null} adjacency matrices for edge (practical) equivalence (one for each group contrast)
 \item  \code{type} \code{ci} (used internally for summarizing and plotting)
 \item  \code{prob} \code{NULL} (used internally for summarizing and plotting)
 }
}
\description{
Edge Differences and (Practical) Equivalence Between GGMs
}
\examples{
# Assume null is true for all edges
Y1 <- MASS::mvrnorm(500, rep(0, 16), Sigma = diag(16))
Y2 <- MASS::mvrnorm(500, rep(0, 16), Sigma = diag(16))
Y3 <- MASS::mvrnorm(500, rep(0, 16), Sigma = diag(16))

ggm_est <- ggm_compare_estimate(Y1, Y2, Y3, iter = 10000, rope = .1)
sel <- select(ggm_est, type = "rope", prob = .95)

# practically equivalent
sel$mats_null
}
