% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.ggm_compare_bf.R
\name{select.ggm_compare_explore}
\alias{select.ggm_compare_explore}
\title{Select Graph: \code{ggm_compare_explore} Objects}
\usage{
\method{select}{ggm_compare_explore}(object, post_prob = 0.5, BF = NULL, ...)
}
\arguments{
\item{object}{object of class \code{ggm_compare_bf}.}

\item{post_prob}{numeric. Posterior `inclusion` probability.  The default is set to 0.50.}

\item{...}{not currently used.}
}
\value{
list of class \code{select.ggm_compare_bf}

\itemize{
\item \code{BF_10} Bayes factors for the alternative ("not equal")
\item \code{BF_01} Bayes factors for the null hypothesis
\item \code{BF_10_adj} Bayes factor adjacency matrix for the alternative ("not equal")
\item \code{BF_01_adj} Bayes factor adjacency matrix for the null hypothesis
\item \code{adj_10} adjacency matrix for the selected edges (in favor of the "not equal")
\item \code{adj_01} adjacency matrix for the selected edges (in favor of the null hypothesis)
\item \code{pcor_mat_10} partial correlation matrix for the alternative ("not equal")
}
}
\description{
Select Graph: \code{ggm_compare_explore} Objects
}
\examples{
\donttest{
Y1 <- MASS::mvrnorm(500, rep(0, 16), Sigma = diag(16))
Y2 <- MASS::mvrnorm(500, rep(0, 16), Sigma = diag(16))
Y3 <- MASS::mvrnorm(500, rep(0, 16), Sigma = diag(16))

ggm_bf <- ggm_compare_bf(Y1, Y2, Y3,
                        prior_sd = .20,
                        iter = 500,
                        cores = 2)
# select with BF_cut = 3
ggm_bf_sel <- select(ggm_bf, BF_cut = 3)

# summary
summary(ggm_bf_sel)
}
}
