% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.estimate.R
\name{select}
\alias{select}
\title{S3 select method}
\usage{
select(object, ...)
}
\arguments{
\item{object}{object of class \code{estimate}, \code{explore}, or ..}

\item{...}{not currently used}
}
\value{
\code{select} works with the following methods:
\itemize{
\item \code{\link{select.estimate}}
\item \code{\link{select.explore}}
\item \code{\link{select.ggm_compare_estimate}}
}
}
\description{
S3 select method
}
