% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.var_estimate.R
\name{select.var_estimate}
\alias{select.var_estimate}
\title{Graph Selection for \code{VAR.estimate} Object}
\usage{
\method{select}{var_estimate}(object, cred = 0.95, alternative = "two.sided")
}
\arguments{
\item{object}{An object of class \code{VAR.estimate}.}

\item{cred}{Numeric. The credible interval width for selecting the graph
(defaults to 0.95; must be between 0 and 1).}

\item{alternative}{A character string specifying the alternative hypothesis. It
must be one of "two.sided" (default), "greater"  or "less".
See note for futher details.}
}
\value{
An object of class \code{select.VAR_estimate}
}
\description{
Graph Selection for \code{VAR.estimate} Object
}
