% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_bggm.explore.R
\name{shiny_bggm}
\alias{shiny_bggm}
\title{\code{shiny_bggm} Generic}
\usage{
shiny_bggm(object, ...)
}
\arguments{
\item{object}{object of class \code{explore}}

\item{...}{currently ignored}
}
\description{
\code{shiny_bggm} Generic
}
