% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.estimate.R
\name{summary.coef.estimate}
\alias{summary.coef.estimate}
\title{Summary method for \code{coef.estimate} objects}
\usage{
\method{summary}{coef.estimate}(object, ...)
}
\arguments{
\item{object}{An object of class \code{coef.estimate}}

\item{...}{currently ignored}
}
\description{
Summary method for \code{coef.estimate} objects
}
\seealso{
\code{\link{coef.estimate}}
}
