% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.estimate.R
\name{summary.coef}
\alias{summary.coef}
\title{Summarize \code{coef} Objects}
\usage{
\method{summary}{coef}(object, cred = 0.95, ...)
}
\arguments{
\item{object}{An object of class \code{coef}.}

\item{cred}{Numeric. The credible interval width for summarizing the posterior
distributions (defaults to 0.95; must be between 0 and 1).}

\item{...}{Currently ignored}
}
\value{
A list of length \emph{p} including the
        summaries for each multiple regression.
}
\description{
Summarize regression parameters with the posterior mean,
standard deviation, and credible interval.
}
\note{
See \code{\link{coef.estimate}} and \code{\link{coef.explore}} for examples.
}
