% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirm.R
\name{summary.confirm}
\alias{summary.confirm}
\title{Summary Method for \code{confirm} Objects}
\usage{
\method{summary}{confirm}(object, ...)
}
\arguments{
\item{object}{object of class \code{confirm}}

\item{...}{currently ignored}
}
\description{
Summary Method for \code{confirm} Objects
}
