% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm_compare_bf.default.R
\name{summary.ggm_compare_bf}
\alias{summary.ggm_compare_bf}
\title{Summary method for \code{ggm_compare_bf} objects}
\usage{
\method{summary}{ggm_compare_bf}(object, ...)
}
\arguments{
\item{object}{An object of class \code{ggm_compare_bf}}

\item{...}{currently ignored}
}
\value{
A list containing the summarized posterior hypothesis probabilities
}
\description{
Summary method for \code{ggm_compare_bf} objects
}
\seealso{
\code{\link{ggm_compare_bf}}
}
