% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm_compare_bf.default.R
\name{summary.ggm_compare_explore}
\alias{summary.ggm_compare_explore}
\title{Summary Method for \code{ggm_compare_explore} Objects}
\usage{
\method{summary}{ggm_compare_explore}(object, col_names = TRUE, ...)
}
\arguments{
\item{object}{object of class \code{ggm_compare_explore}}

\item{col_names}{logical. Should the summary include the column names (default is \code{TRUE})?
Setting to \code{FALSE} includes the column numbers (e.g., \code{1--2}).}

\item{...}{currently ignored}
}
\value{
An object of class \code{summary.ggm_compare_explore}
}
\description{
Summarize the posterior hypothesis probabilities
}
