% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm_compare_ppc.default.R
\name{summary.ggm_compare_ppc}
\alias{summary.ggm_compare_ppc}
\title{Summary method for \code{ggm_compare_ppc.default} objects}
\usage{
\method{summary}{ggm_compare_ppc}(object, ...)
}
\arguments{
\item{object}{An object of class \code{ggm_compare_ppc}}

\item{...}{currently ignored}
}
\value{
A list containing the summarized results
}
\description{
Summary method for \code{ggm_compare_ppc.default} objects
}
\seealso{
\code{\link{ggm_compare_ppc}}
}
