% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{summary.metric}
\alias{summary.metric}
\title{Summary Method for \code{metric} Objects}
\usage{
\method{summary}{metric}(object, cred = 0.95, ...)
}
\arguments{
\item{object}{object of class \code{metric}}

\item{cred}{credible interval}

\item{...}{currently ignored}
}
\description{
Summary Method for \code{metric} Objects
}
