% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.estimate.R
\name{summary.predict}
\alias{summary.predict}
\title{Summary Method for \code{predict} Objects}
\usage{
\method{summary}{predict}(object, ...)
}
\arguments{
\item{object}{object of class \code{predict}}

\item{...}{currently ignored}
}
\description{
Summary Method for \code{predict} Objects
}
