% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.estimate.R
\name{summary.select.estimate}
\alias{summary.select.estimate}
\title{Summary method for \code{select.estimate} objects}
\usage{
\method{summary}{select.estimate}(object, summarize = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{select.estimate}}

\item{summarize}{if \code{TRUE} partial correlations and credible intervals are provided}

\item{...}{currently ignored}
}
\description{
Summary method for \code{select.estimate} objects
}
