% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.select.explore.R
\name{summary.select.explore}
\alias{summary.select.explore}
\title{Summarize \code{select.explore}}
\usage{
\method{summary}{select.explore}(x, hyp = "H1", log = TRUE,
  summarize = FALSE, ...)
}
\arguments{
\item{x}{object of class \code{select.explore}}

\item{hyp}{hypothesis to summarize-e.g., \code{"H0"}, \code{"H1"}, etc.}

\item{log}{log tranform the Bayes factors}

\item{summarize}{\code{summarize = TRUE} provides a summary output of the edges, Bayes factor,
and posterior probabilities. default is \code{FALSE}, which outputs the matrix for the respective hypothesis (i.e., \code{hyp})}

\item{...}{\code{digits}}
}
\description{
Summarize \code{select.explore}
}
\examples{

Y <- BGGM::bfi[,1:10]

# sample from posterior
fit_bf <- explore(Y, prior_sd = 0.5,
                  iter = 5000,
                  cores = 2)

# select the graph
E <- select(fit_bf,
           hyp_prob = 0.90,
           alternative = "exhaustive")

# hyp = "H0"
summary(E, hyp = "H0", summarize = F)

}
