% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.select.ggm_compare_bf.R
\name{summary.select.ggm_compare_bf}
\alias{summary.select.ggm_compare_bf}
\title{Summarize \code{select.ggm_compare_bf}}
\usage{
\method{summary}{select.ggm_compare_bf}(x, type = "adj")
}
\arguments{
\item{type}{"adj" for adjacency matrix (0's and 1's). "BF" for matrix of selected Bayes factors}
}
\description{
Summarize \code{select.ggm_compare_bf}
}
