% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.ggm_compare_estimate.R
\name{summary.select.ggm_compare_estimate}
\alias{summary.select.ggm_compare_estimate}
\title{Summary method for \code{select.ggm_compare_estimate} objects}
\usage{
\method{summary}{select.ggm_compare_estimate}(object, ...)
}
\arguments{
\item{object}{An object of class \code{select.ggm_compare_estimate}}

\item{...}{currently ignored}
}
\description{
Summary method for \code{select.ggm_compare_estimate} objects
}
\seealso{
\code{\link{select.ggm_compare_estimate}}
}
