% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_error.R
\docType{data}
\name{testtt}
\alias{test}
\title{Title}
\format{An object of class \code{prediction_error} of length 4.}
\usage{
test
}
\arguments{
\item{fit}{fitted model from \code{bayes_estimate}}

\item{measure}{mae = mean absolute error; mse = mean squared error, kl = KL-divergence}

\item{newdata}{out-of-sample prediction}

\item{selected}{the selected graphical structure}

\item{ci_width}{credible interval to summarise measure}

\item{samples}{number of posterior samples used}
}
\value{
summary as data frame that can be used for plotting with \code{predictive plot}
}
\description{
Title
}
\examples{

# fit the model
fit <- bayes_estimate(X)

# model selection
select <- estimate_select(fit, ci_width = .99)

# mean squared error
pred_error <- prediction_error(fit, measure = "mse", selected = select$adjacency_mat)

# summary information
pred_error$summary_error

# plot
plt <- predictive_plot(pred_error, size = 2, color = "blue")
}
\keyword{datasets}
