% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{women_math}
\alias{women_math}
\title{Data: Women and Mathematics}
\format{
A data frame containing 1190 observations (n = 1190) and 6 variables (p = 6) measured on the binary scale
        \insertCite{fowlkes1988evaluating}{BGGM}. These data have been analyzed in \insertCite{tarantola2004mcmc;textual}{BGGM}
        and in \insertCite{madigan1994model}{BGGM}. The variable descriptions were copied from  (section 5.2 )
        \insertCite{@section 5.2, @talhouk2012efficient}{BGGM}
}
\usage{
data("women_math")
}
\description{
A data frame containing 1190 observations (n = 1190) and 6 variables (p = 6) measured on the binary scale.
}
\details{
\itemize{
  \item \code{1}  Lecture attendance (attend/did not attend)
  \item \code{2}  Gender (male/female)
  \item \code{3}  School type (urban/suburban)
  \item \code{4}  “I will be needing Mathematics in my future work” (agree/disagree)
  \item \code{5}  Subject preference (math/science vs. liberal arts)
  \item \code{6} Future plans (college/job)
}
}
\examples{
data("women_math")
}
\references{
\insertAllCited{}
}
\keyword{datasets}
