% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictability.R
\name{assess_predictability}
\alias{assess_predictability}
\title{Assess Predictability of \code{Predictability} Objects}
\usage{
assess_predictability(..., contrast = "exhaustive", prob = 0.95)
}
\arguments{
\item{...}{object(s) of class \code{predictability}}

\item{prob}{}
}
\value{

}
\description{
Assess Predictability of \code{Predictability} Objects
}
