% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.estimate.R
\name{coef.estimate}
\alias{coef.estimate}
\title{Compute Regression Parameters}
\usage{
\method{coef}{estimate}(object, iter = NULL, ...)
}
\arguments{
\item{object}{object of class \code{estimate} or \code{explore}}

\item{iter}{number of iterations (posterior samples; defaults to the number in the object).}

\item{...}{e.g., \code{digits}}
}
\value{
list of class \code{coef.estimate}:
}
\description{
Compute Regression Parameters
}
\examples{
# p = 10
Y <- BGGM::bfi[,1:10]

# sample posterior
fit <- estimate(Y, iter = 5000)

# precision to regression
coefficients(fit, node = 1, cred = 0.95)
}
