% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.estimate.R
\name{coef.explore}
\alias{coef.explore}
\title{Compute Regression Parameters for \code{explore} Objects}
\usage{
\method{coef}{explore}(object, iter = NULL, ...)
}
\arguments{
\item{object}{An Object of class \code{explore}.}

\item{iter}{Number of iterations (posterior samples; defaults to the number in the object).}

\item{...}{Currently ignored.}
}
\value{
An object of class \code{coef}, containting two lists.

\itemize{

\item \code{betas} A list of length \emph{p}, each containing a \emph{p} - 1 by \code{iter} matrix of
posterior samples

\item \code{object} An object of class \code{explore} (the fitted model).
}
}
\description{
There is a direct correspondence between the inverse covariance matrix and
 multiple regression \insertCite{kwan2014regression,Stephens1998}{BGGM}. This readily allows
 for converting the GGM paramters to regression coefficients. All data types are supported.
}
\examples{
\donttest{
# note: iter = 250 for demonstrative purposes
# data
Y <- ptsd[,1:4]

##########################
### example 1: ordinal ###
##########################

# fit model (note + 1, due to zeros)
fit <- explore(Y + 1,
               type = "ordinal",
               iter = 250,
               progress = FALSE)

# summarize the partial correlations
reg <- coef(fit)

# summary
summ <- summary(reg)

summ
}
}
\references{
\insertAllCited{}
}
