% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.estimate.R
\name{coef.explore}
\alias{coef.explore}
\title{Compute Regression Parameters for \code{explore} Objects}
\usage{
\method{coef}{explore}(object, iter = NULL, ...)
}
\arguments{
\item{object}{An Object of class \code{estimate}}

\item{iter}{Number of iterations (posterior samples; defaults to the number in the object).}

\item{...}{Not currently used.}
}
\value{
An object of class \code{coef}, containting two lists.

\itemize{

\item \code{betas} A list of length \emph{p}, each containing a \emph{p} - 1 by \code{iter} matrix of
posterior samples

\item \code{object} An object of class \code{explore} (the fitted model).
}
}
\description{
There is a direct correspondence between the inverse covariance matrix and
 multiple regression \insertCite{kwan2014regression,Stephens1998}{BGGM}. This readily allows
 for converting the GGM paramters to regression coefficients. All data types are supported.
}
\references{
\insertAllCited{}
}
