% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirm.R
\name{confirm}
\alias{confirm}
\title{Confirmatory Hypothesis Testing in GGMs}
\usage{
confirm(
  Y,
  hypothesis,
  prior_sd = 0.25,
  formula = NULL,
  data = NULL,
  type = "continuous",
  iter = 25000,
  cores = 2
)
}
\arguments{
\item{Y}{matrix (or data frame) of dimensions \emph{n} (observations) by  \emph{p} (variables).}

\item{hypothesis}{character string. hypothesis (or hypotheses) to be tested. See notes for futher details.}

\item{prior_sd}{hypothesized standard deviation of the prior distribution}

\item{formula}{an object of class \code{\link[stats]{formula}}. This allows for including
control variables in the model (e.g.,, \code{~ gender * education}).}

\item{data}{an optional data frame, list or environment (or an object coercible by \code{\link[base]{as.data.frame}})
to a data frame containing the variables in \code{formula}. This is required when controlling for variables.}

\item{type}{character string. Which type of data for \strong{Y} ? The options include \code{continuous},
\code{binary}, \code{ordinal}, or \code{mixed}. See the note for further details.}

\item{iter}{posterior and prior samples. 25,000 is the default, as it results in a more stable Bayes factor than
using, say, 5,000.}

\item{cores}{integer. How many cores for parallel computing ? (default is 2).}
}
\value{
list of class \code{confirm}:

\itemize{
\item \code{BF_matrix} matrix of Bayes factors for each hypothesis. Also includes the compliment
\item \code{post_prob} posterior hypothesis probabilities
\item \code{hypotheses} \code{hypothesis}
\item \code{call} match.call()
\item \code{p} number of variables
\item \code{n} number of observations
\item \code{iter} number of posterior samples
\item \code{delta} hyperparameter of matrix-F prior distribution (corresponds to prior_sd)
\item \code{parcors_mat} partial correlation matrix
\item \code{returned_mats} contrast matrices
}
}
\description{
Confirmatory hypothesis testing in GGMs, for, say, comparing formalized models encoding
scienctific expectations. Hypotheses are expressed as equality and/or ineqaulity contraints on the
partial correlations of interest. Here the focus is \emph{not} on determining the graph (see \code{\link{explore}}) but testing specific hypotheses related to
the conditional (in)dependence structure. These methods were introduced in \insertCite{Williams2019_bf;textual}{BGGM}.
}
\note{
\strong{Interpretation of conditional (in)dependence models for latent data:}

A  tetrachoric correlation (binary data) is a special case of a polychoric correlation (ordinal data). Both relations are
between "theorized normally distributed continuous latent variables"
(\href{https://en.wikipedia.org/wiki/Polychoric_correlation}{Wikipedia})
In both instances, the correpsonding partial correlation between observed variables is conditioned
on the remaining variables in the \emph{latent} space. This implies that interpration is much the same as
for continuous data, but with respect to latent variables. We refer interested reader to
\insertCite{@page 2364, section 2.2, in  @webb2008bayesian;textual}{BGGM}.


\strong{Hypothesis Examples:}

The hypotheses can be written either with the respective column names or numbers. For example, \code{1--2} denotes the
relation between variables in column 1 and 2. Note that these must correspond to the upper triangular elements of the correlation
matrix. This is accomplished by ensuring that the first number is smaller than the second number. This also applies when using
column names (e.g., in reference to the column number).

One hypothesis:
\itemize{

\item  To test whether some relations are larger than others, while others are expected to be equal, this can be writting as

 \code{hyp <-  c(1--2 > 1--3  = 1--4 > 0)},

 where there is an addition additional contraint that all effects are expected to be positive. This is then compared to the complement.
}

More than one hypothesis:
\itemize{

\item The above hypothesis can also be compared to, say, a null model by using ";" to seperate the hypotheses, for example,

\code{hyp <-  c(1--2 > 1--3  = 1--4 > 0; 1--2 = 1--3  = 1--4 = 0)}.

Any number of hypotheses can be tested.
}



see \code{methods(class = "confirm")}
}
\examples{
\donttest{

# p = 10
Y <- BGGM::bfi[,1:10]

# hypothesis
hypothesis <- c("1--2 > 1--3 > 1--4 > 1--5")

# test inequality contraint
test_order <-  confirm(Y = Y, hypothesis  = hypothesis,
                      prior_sd = 0.5, iter = 50000,
                      cores = 2)
# summary
summary(test_order)


# test hypothesized directions

# hypothesis
hypothesis <- c("(1--2, 1--3, 1--4)  <  0 < (1--6)")

# test directions
test_directions <-  confirm(Y = Y, hypothesis  = hypothesis,
                      prior_sd = 0.5, iter = 50000,
                      cores = 2)
# summary
test_directions

#############################
######## Binary #############
#############################




}
}
\references{
\insertAllCited{}
}
