% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_compare.estimate.R
\name{edge_compare.estimate}
\alias{edge_compare.estimate}
\title{Title}
\usage{
\method{edge_compare}{estimate}(x, contrast, ci_width, rope = NULL)
}
\arguments{
\item{rope}{}
}
\description{
Title
}
