% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate.default}
\alias{estimate.default}
\title{estimate}
\usage{
\method{estimate}{default}(x, samples = 5000, ...)
}
\arguments{
\item{x}{data matrix}

\item{samples}{number of posterior samples}
}
\value{
An object of class \code{estimate} that includes posterior samples, partial correlation matrix, etc. Note
this object is then used with other functions for model selection, prediction, regression coverion, etc.
For Bayesian hypothesis testing see \code{explore} and \code{confirm}.
}
\description{
Samples from the posterior with the Wishart prior distribution.
}
\examples{
fit <- estimate(X)
fit
}
