% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{S3 estimate method}
\usage{
estimate(Y, iter = 500, analytic = FALSE, ...)
}
\arguments{
\item{Y}{data matrix (\emph{n} by  \emph{p}).}

\item{iter}{number of posterior samples}

\item{analytic}{analytic solution. see notes for further details.}

\item{...}{currently not used}
}
\description{
S3 estimate method
}
\seealso{
\code{\link{estimate.default}}
}
