% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_select.R
\name{estimate_select}
\alias{estimate_select}
\title{Title}
\usage{
estimate_select(x, ci_width)
}
\arguments{
\item{x}{object of class \code{bayes_estimate}}

\item{ci_width}{credible interval width used for the decision rule}
}
\value{
he selected partial correlation matrix, that is the *conditional* dependence structure,
is returned along with the adjacency matrix
}
\description{
Title
}
