% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_ordinal.R
\name{gen_ordinal}
\alias{gen_ordinal}
\title{Generate Ordinal and Binary data}
\usage{
gen_ordinal(n, p, levels, cor_mat, empirical = FALSE)
}
\arguments{
\item{n}{number of observations}

\item{p}{number of variables}

\item{levels}{number of categories}

\item{cor_mat}{true correlation matrix}

\item{empirical}{logical. If true, mu and Sigma specify the empirical not
population mean and covariance matrix.}
}
\value{

}
\description{
Generate Ordinal and Binary data
}
