% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGM_compare_bf.R
\name{GGM_compare_bf}
\alias{GGM_compare_bf}
\title{GGM_compare_bf}
\usage{
GGM_compare_bf(Y_g1, Y_g2, Y_g3 = NULL, Y_g4 = NULL, hyp, priorprob = 1,
  cores, delta, nu, n_samples)
}
\arguments{
\item{n_samples}{}
}
\description{
GGM_compare_bf
}
