% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm_compare_bf.default.R
\name{ggm_compare_bf}
\alias{ggm_compare_bf}
\title{S3 estimate method}
\usage{
ggm_compare_bf(...)
}
\arguments{
\item{...}{currently not used}
}
\description{
S3 estimate method
}
\seealso{
\code{\link{ggm_compare_bf.default}}
}
