% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm_compare_estimate.default.R
\name{ggm_compare_estimate.default}
\alias{ggm_compare_estimate.default}
\title{Compare Edges Between GGMs with the Posterior Distribution}
\usage{
\method{ggm_compare_estimate}{default}(..., iter = 5000,
  ci_width = NULL, rope = NULL)
}
\arguments{
\item{...}{data matrices (\emph{n} \times  \emph{p}). Requires at least two.}

\item{iter}{number of posterior samples}

\item{ci_width}{credible interval width used for the decision rule}

\item{rope}{region of practical equivalence}
}
\description{
Compare Edges Between GGMs with the Posterior Distribution
}
