% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm_compare_estimate.default.R
\name{ggm_compare_estimate}
\alias{ggm_compare_estimate}
\title{GGM Compare: Estimate}
\usage{
ggm_compare_estimate(
  ...,
  formula = NULL,
  data = NULL,
  type = "continuous",
  mixed_type = NULL,
  analytic = FALSE,
  prior_sd = 0.25,
  iter = 5000,
  seed = 1
)
}
\arguments{
\item{...}{matrices (or data frame) of dimensions \emph{n} (observations) by  \emph{p} (variables).
Requires at least two.}

\item{formula}{an object of class \code{\link[stats]{formula}}. This allows for including
control variables in the model (i.e., \code{~ gender}).}

\item{data}{an optional data frame, list or environment (or an object coercible by \code{\link[base]{as.data.frame}})
to a data frame containing the variables in \code{formula}. This is required when controlling for variables.}

\item{type}{character string. Which type of data for \strong{Y} ? The options include \code{continuous},
\code{binary}, or \code{ordinal}. See the note for further details.}

\item{mixed_type}{numeric vector. An indicator of length p for which varibles should be treated as ranks.
(1 for rank and 0 to assume normality). The default is currently (dev version) to treat all integer variables
as ranks when \code{type = "mixed"} and \code{NULL} otherwise. See note for further details.}

\item{analytic}{logical. Should the analytic solution be computed (default is \code{FALSE}) ?}

\item{prior_sd}{The scale of the prior distribution (centered at zero), in reference to a beta distribtuion.
The `default` is 0.25. See note for further details.}

\item{iter}{number of iterations (posterior samples; defaults to 5000).}

\item{seed}{The seed for random number generation (default set to \code{1}).}
}
\value{
A list of class \code{ggm_compare_estimate} containing:
 \itemize{
 \item \code{pcor_diffs} partial correlation differences (posterior distribution)
 \item \code{p} number of variable
 \item \code{info} list containing information about each group (e.g., sample size, etc.)
 \item \code{iter} number of posterior samples
 \item \code{call} \code{match.call}
 }
}
\description{
Compare edges (partial correlations) that are estimated from groups to, say, detect a differences or equivalence.
}
\note{
The work flow for most functions in \strong{BGGM} is to first fit the model
and then select the graph  (in this case the differences) with  \code{\link{select}}.
}
\examples{
# data
Y1 <- BGGM::bfi[1:500,1:5]
Y2 <- BGGM::bfi[501:1000, 1:5]

# fit model
fit <- ggm_compare_estimate(Y1, Y2)

# posterior summary of differences
summary(fit)

# select (threshold) with credible intervals
sel <- select(fit)

# summary
summary(sel)

# selected differences
sel$mat_pcor

# adjacency matrix
sel$mat_adj
}
\seealso{
\code{\link{select.ggm_compare_estimate}}
}
