% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm_compare_bf.default.R
\name{ggm_compare_explore}
\alias{ggm_compare_explore}
\title{GGM Compare: Exploratory Hypothesis Testing}
\usage{
ggm_compare_explore(
  ...,
  formula = NULL,
  type = "continuous",
  mixed_type = NULL,
  analytic = FALSE,
  prior_sd = 0.2,
  iter = 5000
)
}
\arguments{
\item{...}{at least two matrices (or data frame) of dimensions \emph{n} (observations) by  \emph{p} (variables).}

\item{formula}{an object of class \code{\link[stats]{formula}}. This allows for including
control variables in the model (i.e., \code{~ gender}).}

\item{type}{character string. Which type of data for \strong{Y} ? The options include \code{continuous},
\code{binary}, or \code{ordinal}. See the note for further details.}

\item{mixed_type}{numeric vector. An indicator of length p for which varibles should be treated as ranks.
(1 for rank and 0 to assume normality). The default is currently (dev version) to treat all integer variables
as ranks when \code{type = "mixed"} and \code{NULL} otherwise. See note for further details.}

\item{analytic}{logical. Should the analytic solution be computed (default is \code{FALSE}) ? See note for details.}

\item{prior_sd}{The scale of the prior distribution (centered at zero), in reference to a beta distribtuion.
The `default` is 0.25. See note for further details.}

\item{iter}{number of iterations (posterior samples; defaults to 5000).}

\item{data}{an optional data frame, list or environment (or an object coercible by \code{\link[base]{as.data.frame}})
to a data frame containing the variables in \code{formula}. This is required when controlling for variables.}

\item{seed}{The seed for random number generation (default set to \code{1}).}
}
\value{
list of class \code{ggm_compare_bf}
\itemize{
\item \code{BF_01}  Bayes factors in favor of the null hypothesis.
\item \code{pcor_diff} partial correlation difference (for two groups only)
\item \code{p} number of variables
\item  \code{info} list of information about the data matrices

\itemize{
\item \code{dat} list containing the data matrices
\item \code{dat_info} sample size for each data matrix
\item \code{pairwise} matrix of pairwise combinations
}
\item \code{iter} number of posterior and prior samples
\item \code{call} match.call()
\item \code{delta} hyperparameter of matrix-F prior distribution (corresponds to \code{prior_sd})

\item \code{groups} number of groups
\item \code{post_samp} matrix of posterior samples
}
}
\description{
Compare Gaussian graphical models with (exploratory) Bayesian hypothesis testing using the matrix-F prior
distribution \insertCite{Mulder2018}{BGGM}. A test for each partial correlation in the model for any number of groups.
This provides evidence for the null hypothesis of no difference and the alternative hypothesis
of difference. With more than two groups, the test is for \emph{all} groups simultaneously (i.e., the relation is the same or different in all groups).
This method was introduced in \insertCite{williams2020comparing;textual}{BGGM}. For confirmatory hypothesis testing see
\code{confirm_groups}.
}
\note{
\strong{More than Two Groups:}

 It is possible to compare any number of group. This does \emph{not} provide pairwise differences. Rather, the test
 is for all group, say, that each partial correlation is the same in each. This is analagous to an ANOVA--an
 `omnibus` test that does not indicate which groups are different. A follow up test would be required for this purpose.

\strong{A Note on Defaults:}

The `default` for \code{prior_sd} is 0.25. This can and should be changed to reflect the hypothesized difference.
If differences are expected to be small \code{prior_sd} should be set to a smaller value (e.g., 0.15).
This might raise concerns of the `correct` prior.` Note, however, that the interpretation remains unchanged: which hypothesis better
predicts the observed data. \insertCite{@p. 777, in  @Kass1995}{BGGM}

If the desired inference is \emph{not} (relative) evidence between models, the method in \code{\link{ggm_compare_estimate}}
(for each partial correlation ) or  \code{\link{ggm_compare_ppc}} (a global test) can be used.


\strong{Interpretation of conditional (in)dependence models for latent data:}

A  tetrachoric correlation (binary data) is a special case of a polychoric correlation (ordinal data). Both relations are
between "theorized normally distributed continuous latent variables"
(\href{https://en.wikipedia.org/wiki/Polychoric_correlation}{Wikipedia})
In both instances, the correpsonding partial correlation between observed variables is conditioned
on the remaining variables in the \emph{latent} space. This implies that interpration is much the same as
for continuous data, but with respect to latent variables. We refer interested reader to
\insertCite{@page 2364, section 2.2, in  @webb2008bayesian;textual}{BGGM}.


\strong{Mixed Data:}
}
\examples{
\donttest{
# assume null is true
Y1 <- MASS::mvrnorm(500, rep(0, 16), Sigma = diag(16))
Y2 <- MASS::mvrnorm(500, rep(0, 16), Sigma = diag(16))
Y3 <- MASS::mvrnorm(500, rep(0, 16), Sigma = diag(16))

# fit model
bf_ggm <- ggm_compare_bf(Y1, Y2, Y3,
                         prior_sd = .2,
                         iter = 500, cores = 2)

# select graph
sel <- select(bf_ggm, BF_cut = 3)

# plot
plot(sel)

}
}
\references{
\insertAllCited{}
}
