% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGM_compare_ppc.R
\name{GGM_compare_ppc}
\alias{GGM_compare_ppc}
\title{GGM comparison based on the predictive distribution}
\usage{
GGM_compare_ppc(Y_g1, Y_g2, Y_g3 = NULL, Y_g4 = NULL, contrast = NULL,
  type = NULL, compare_all = NULL, samples)
}
\arguments{
\item{samples}{numer of posterior and predictive samples}

\item{groups}{}
}
\description{
GGM comparison based on the predictive distribution
}
