% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_plot.R
\name{hypothesis_plot}
\alias{hypothesis_plot}
\title{Hypothesis Plot}
\usage{
hypothesis_plot(rho_sd = NULL, difference = FALSE, fit = NULL,
  edge = NULL, samples = 10000, size = NULL)
}
\arguments{
\item{rho_sd}{standard deviation of the prior distribution (see notes)}

\item{difference}{\code{TRUE} plots the difference between prior distributions (see notes)}

\item{fit}{option object of class \code{explore}}

\item{edge}{option for plotting selected posterior distribtion (only if a fitted model is provided)}

\item{samples}{number of samples drawn from the prior distirbition}

\item{size}{point size for prior and posteriro plot (only needed if a fitted model is provided)}
}
\value{
a \code{ggplot2} object
}
\description{
This function allows for plotting prior distributions and prior distribution differences.
 This can assist in selecting a hypothesized effect size, which is used for computing the the Bayes factor. Further,
 if a fitted object is supplied (\code{fit}), this allows for plotting the chosen prior and a selected posterior distribution (\code{edge}).
 This shows how the Bayes factor is computed (ratio of denisty evaluated at zero).
}
\examples{

# plot prior distributions
plot_priors <- hypothesis_plot(c(0.25, 0.5))
plot_prior

#plot prior distribution differences
plot_priors <- hypothesis_plot(c(0.25, 0.5))
plot_prior

# plot posterior and prior
Y <- BGGM::ptsd[,1:10]
fit <- explore(Y, prior_sd = .5, iter = 5000)

plot_post_prior <- hypothesis_plot(fit = fit, edge = "2--3", size = 4, difference = FALSE)
plot_post_prior

}
