% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lincomb.R
\name{lincomb}
\alias{lincomb}
\title{Test a linear combination of posterior samples}
\usage{
lincomb(lin_comb, object, cred = 0.95, rope = NULL)
}
\arguments{
\item{lin_comb}{A string specifying a linear combination of variables}

\item{object}{An object of class \code{BGGM}}

\item{cred}{Numeric. The credible interval width
(defaults to 0.95; must be between 0 and 1).}

\item{rope}{Specify a region of practical equivalence (i.e., ROPE). This is Optional.}
}
\value{
An object of class \code{lin_comb}
}
\description{
Test a linear combination of posterior samples
}
\examples{
\donttest{
# data
Y <- BGGM::ptsd

# names
colnames(Y) <- letters[1:20]

# estimate model
est <- BGGM::estimate(Y)

# test
bggm_comb <- lincomb("a--c + a--d > b--c + b--d",
                       object = est,
                       cred = 0.90)

# print
bggm_comb
}
}
