% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\title{Maximum A Posteriori Precision Matrix}
\usage{
map(Y)
}
\arguments{
\item{Y}{Matrix (or data frame) of dimensions \emph{n} (observations) by  \emph{p} (variables).}
}
\value{
An object of class \code{map}, including the precision matrix,
        partial correlation matrix, and regression parameters.
}
\description{
Maximum A Posteriori Precision Matrix
}
\examples{
Y <- BGGM::bfi[, 1:5]

# map
map <- map(Y)
map
}
