% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{mape}
\alias{mape}
\title{Mean Absolute Percentage Error for \code{predict.estimate} Objects}
\usage{
mape(object, ...)
}
\arguments{
\item{object}{object of class \code{post.pred} or  \code{predict.estimate}}

\item{...}{currently ignored}
}
\value{

}
\description{
Mean Absolute Percentage Error for \code{predict.estimate} Objects
}
\examples{
# data
Y <- subset(tas, gender == "M")[,-ncol(tas)]

# fit model
fit <- estimate(Y)

# predict (note summary = FALSE)
pred <- predict(fit, summary = FALSE)

mape(pred)
}
