% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_imputation.R
\name{mvn_imputation}
\alias{mvn_imputation}
\title{Multivariate Normal Imputation}
\usage{
mvn_imputation(Y, type = "continuous", iter = 1000, progress = TRUE)
}
\arguments{
\item{Y}{Matrix (or data frame) of dimensions \emph{n} (observations) by  \emph{p} (variables).}

\item{type}{Character string. Which type of data for \code{Y} ? The options include \code{continuous},
\code{binary}, \code{ordinal}, or \code{mixed}. Note that mixed can be used for data with only
ordinal variables. See the note for further details.}

\item{iter}{Number of iterations (posterior samples; defaults to 5000).}

\item{progress}{Logical. Should a progress bar be included (defaults to \code{TRUE}) ?}
}
\value{
An object of class \code{mvn_imputation}
}
\description{
Impute values, assuming a multivariate normal distribution, with the posterior
predictive distribution.
}
\examples{
\donttest{

d <- scale(na.omit(subset(ifit, id ==1)[,-1]))[,1:7]

Y <- MASS::mvrnorm(5000, rep(0, 7), cor(d))

# for checking
Ymain <- Y

indices <- which( matrix(0, 5000, 7) == 0, arr.ind = TRUE)
na_indices <- indices[sample(5:nrow(indices), size = 1000, replace = F),]
na_indices <- na_indices[na_indices[,1] > 1,]

Y[na_indices] <- NA
Y_miss <- ifelse(is.na(Y), 1, 0)

# true values
true <- unlist(sapply(1:7, function(x)  Ymain[ which(Y_miss[,x] == 1),x] ))

# impute
fit_missing <- mvn_imputation(Y, progress = FALSE)

# correlation
cor( true, colMeans(fit_missing$ppc_missing))

}
}
