% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm_compare_ppc.default.R
\name{plot.ggm_compare_ppc}
\alias{plot.ggm_compare_ppc}
\title{Plot \code{ggm_compare_ppc} Objects}
\usage{
\method{plot}{ggm_compare_ppc}(
  x,
  critical = 0.05,
  col_noncritical = "#84e184A0",
  col_critical = "red",
  point_size = 2,
  ...
)
}
\arguments{
\item{x}{An object of class \code{ggm_compare_ppc}}

\item{critical}{'Significance' level (defaults to 0.05).}

\item{col_noncritical}{Fill color for the non critical region as a
character (defaults to "#84e184A0").}

\item{col_critical}{Fill color for the critical region as a character.}

\item{point_size}{Numeric point size for the observed score.}

\item{...}{Currently ignored}
}
\value{
An object (or list of objects) of class \code{ggplot}
}
\description{
Plot the predictive check with
\href{https://CRAN.R-project.org/package=ggridges/vignettes/introduction.html}{ggridges}.
}
