% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.predict.R
\name{plot.predict}
\alias{plot.predict}
\title{Predictive Plots for In- and Out-of-Sample Data}
\usage{
\method{plot}{predict}(x, x2 = NULL, size = 2, color = "red",
  width = 0.1, order = NULL, ...)
}
\arguments{
\item{x}{object of class \code{predict}}

\item{x2}{optional object of class \code{predict} (typically using \code{test_data})}

\item{size}{geom_point size}

\item{color}{geom_point color}

\item{width}{geom_errorbar width}

\item{order}{plotting order for the nodes (\code{"train"}
or \code{"test"} error)}

\item{...}{currently ignored}
}
\value{
A baseline \code{ggplot} object. Further customization is possible. An example is provided below.
}
\description{
Predictive Plots for In- and Out-of-Sample Data
}
\note{
x1 can be a \code{predict} object that assesses predictive accuracy on in- or out-of-sample (\code{test_data}) data. x2 can be used to
visually compare in- and out-of sample predictive accuracy. This is accomplished by providing the respective \code{predict} objects to
x1 and x2. An example is provided below.
}
\examples{

dat <- BGGM::ptsd
fit <- estimate(dat, iter = 5000)

error <- predict(fit,
              measure = "R2",
              test_data  = NULL,
              ci_width = 0.95,
              samples = 1000)

plt <- plot(error, size = 5, color = "red", width = 1)

# plot can be changed further with ggplot

plt
}
