% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.predict.R
\name{plot.predict}
\alias{plot.predict}
\title{Predict plot}
\usage{
\method{plot}{predict}(x, size = 2, color = "red", width = 0.1)
}
\arguments{
\item{x}{object of class "predict"}

\item{size}{geom_point size}

\item{color}{geom_point color}

\item{width}{geom_errorbar width}
}
\value{
ggplot object
}
\description{
Predict plot
}
\examples{

dat <- BGGM::ptsd
fit <- estimate(dat, samples)

select graph
selected <- select(fit, ci_width = 0.95)$adjacency_mat

error <- predict(fit,
              selected = selected,
              measure = "MSE",
              test_data  = NULL,
              ci_width = 0.95,
              samples = 1000)

plt <- plot(error, size = 5, color = "red", width = 1)

# plot can be changed further with ggplot--e.g., changing theme

plt + theme_classic()
}
