% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.select.explore.R
\name{plot.select.explore}
\alias{plot.select.explore}
\title{Plot Exploratory (hypothesis testing) Based Graphs}
\usage{
\method{plot}{select.explore}(x, type, lower_tri = TRUE,
  layout = "circle", node_outer = 10, node_inner = 9,
  node_text_size = 8)
}
\arguments{
\item{x}{object of class \code{select.estimate}}

\item{type}{\code{heatmap} or \code{network} plot}

\item{lower_tri}{heatmap lower triangular elements}

\item{layout}{layout of network plot. see igraph for options}

\item{node_outer}{border for node}

\item{node_inner}{node size (see notes)}

\item{node_text_size}{node text size}
}
\description{
Allows for plotting graphs selected wtih Bayesian hypothesis testing. Options include a heatmap or network plot.
The conditional independence and dependence structure are visualized in seperate plots.
}
\examples{

# p =  10
Y <- BGGM::bfi[,1:10]

# sample from posterior
fit_bf <- explore(Y, prior_sd = 0.5,
                 iter = 5000,
                 cores = 2)

# rho > 0
E_pos <- select(fit_bf,
               BF_cut = 3,
               alternative = "greater")

# positive plot
plt_pos <- plot(E_pos, type = "network")
plt_pos <- plt_pos$plot_nonzero +
ggtitle(expression(atop(H[0]: rho[i][j]*" = "*0, H[1]: rho[i][j]*" > "*0)))

plt_pos

}
