% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.select.ggm_compare_bf.R
\name{plot.select.ggm_compare_bf}
\alias{plot.select.ggm_compare_bf}
\title{Plot Edge Differences and Equivalence Between GGMs (\code{ggm_compare_bf })}
\usage{
\method{plot}{select.ggm_compare_bf}(x, mat_type = "adj",
  type = "heatmap", lower_tri = TRUE, layout = "circle",
  node_outer = 10, node_inner = 9, node_text_size = 8,
  labels = NULL)
}
\arguments{
\item{x}{object of class \code{select.ggm_compare_bf}}

\item{mat_type}{"adj" or "BF"}

\item{type}{\code{heatmap} or \code{network} plot (network is currently the only option)}

\item{lower_tri}{heatmap lower triangular elements}

\item{layout}{layout of network plot. see igraph for options}

\item{node_outer}{border for node}

\item{node_inner}{node size (see notes)}

\item{node_text_size}{node text size}

\item{labels}{node labels. Default is 1:p, but can change to colnames(Y)}
}
\value{
objects of class \code{ggplot}
}
\description{
Plot Edge Differences and Equivalence Between GGMs (\code{ggm_compare_bf })
}
\examples{
# data (null is true)
Y1 <- MASS::mvrnorm(500, mu = rep(0,16),
                    Sigma = Sigma = BGGM::ptsd_cor2,
                    empirical = T)
Y2 <- MASS::mvrnorm(500, mu = rep(0,16),
                    Sigma = BGGM::ptsd_cor3,
                    empirical = T)
# fit model
fit <- ggm_compare_bf(Y1, Y2,
                     prior_sd = 0.22,
                     iter = 5000,
                     cores = 4)
# select E
sel <- select(fit, BF_cut = 3)
# plot
plot(sel, mat_type = "BF", type = "heatmap")
}
