% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{plot.summary.estimate}
\alias{plot.summary.estimate}
\title{Plot \code{summary.estimate} Objects}
\usage{
\method{plot}{summary.estimate}(x, color = "black", size = 2, width = 0, ...)
}
\arguments{
\item{x}{an object of class \code{summary.estimate}}

\item{color}{color of error bar}

\item{size}{Numeric. The size for the points.}

\item{width}{width of error bar cap}

\item{...}{currently ignored}
}
\value{
an object of class \code{ggplot}
}
\description{
Plot \code{summary.estimate} Objects
}
