% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.default.R
\name{plot.summary.explore}
\alias{plot.summary.explore}
\title{Plot \code{summary.explore}}
\usage{
\method{plot}{summary.explore}(x, ...)
}
\arguments{
\item{x}{an object of class \code{summary.explore}}

\item{...}{currently ignored}
}
\value{
an object of class \code{ggplot}
}
\description{
Plot \code{summary.explore}
}
