% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm_compare_estimate.default.R
\name{plot.summary.ggm_compare_estimate}
\alias{plot.summary.ggm_compare_estimate}
\title{Plot \code{summary.ggm_compare_estimate} Objects}
\usage{
\method{plot}{summary.ggm_compare_estimate}(x, color = "black", size = 2, width = 0, ...)
}
\arguments{
\item{x}{an object of class \code{estimate} or \code{ggm_compare_estimate}}

\item{color}{color of error bar}

\item{width}{width of error bar cap}

\item{...}{currently ignored}
}
\value{
an object of class \code{ggplot}
}
\description{
Plot \code{summary.ggm_compare_estimate} Objects
}
