% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm_compare_bf.default.R
\name{plot.summary.ggm_compare_explore}
\alias{plot.summary.ggm_compare_explore}
\title{GGM Compare: Plot \code{ggm_compare_explore] Object}}
\usage{
\method{plot}{summary.ggm_compare_explore}(x, size = 2, color = "black")
}
\arguments{
\item{x}{object of class \code{ggm_compare_explore}}

\item{size}{numeric. Size of the points.}

\item{color}{string. Color of the points.}
}
\value{

}
\description{
Visualize posterior hypothesis probabilities.
}
