% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prior.R
\name{plot_prior}
\alias{plot_prior}
\title{Plot: Prior Distribution}
\usage{
plot_prior(prior_sd = 0.2, iter = 5000)
}
\arguments{
\item{delta}{numeric. Hyperparameter that governs the scale of
             the prior distribution. Larger values are more informative
             (i.e., narrower).

 @param iter number of iterations (posterior samples; defaults to 5000).}
}
\value{
a \code{ggplot} object.
}
\description{
Visualize the implied prior distribution for the partial correlations. This is
             particularly useful for the Bayesian hypothesis testing methods.

Visualize the prior distribution used for the
Bayesian hypothesis testing methods. This is governed by \code{prior_sd}.
}
