% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prior.R
\name{plot_prior}
\alias{plot_prior}
\title{Plot: Prior Distribution}
\usage{
plot_prior(prior_sd = 0.5, iter = 5000)
}
\arguments{
\item{prior_sd}{Scale of the prior distribution, approximately the standard deviation
of a beta distribution (defaults to 0.5).}

\item{iter}{Number of iterations (prior samples; defaults to 5000).}
}
\value{
A \code{ggplot} object.
}
\description{
Visualize the implied prior distribution for the partial correlations. This is
             particularly useful for the Bayesian hypothesis testing methods.
}
\examples{
# note: iter = 250 for demonstrative purposes

plot_prior(prior_sd = 0.25, iter = 250)
}
