% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.estimate.R
\name{predict.estimate}
\alias{predict.estimate}
\title{Nodewise In- and Out-of-Sample Predictability}
\usage{
\method{predict}{estimate}(object, test_data = NULL, ci_width = 0.95,
  iter = 1000, measure = "R2", ...)
}
\arguments{
\item{object}{object of class \code{estimate}}

\item{test_data}{optional test data set}

\item{ci_width}{select graph and width of interval in the summary output}

\item{iter}{number of posterior samples used to compute predictability
(must be the same or less than the number of samples in the \code{estimate} object)}

\item{measure}{Bayesian R2 (R2) or mean squared error (MSE)}

\item{...}{currently ignored}
}
\value{
object of class \code{predict}
\itemize{
\item \code{summary_error} summarized predictability
\item \code{poster_samples} list containing the posterior samples for each node
}
}
\description{
Bayesian predictability. The measures are computed with respect to the posterior distributions.
This provides uncertainty for variance explained (Bayesian R2) and mean squared error.
}
\examples{

dat <- BGGM::ptsd
fit <- estimate(dat, iter = 5000)

error <- predict(fit,
              measure = "MSE",
              test_data  = NULL,
              ci_width = 0.95,
              samples = 1000)

summary(error)

plot(error)

################################
## training/test data example ##
################################

train_dat <- BGGM::ptsd[1:200,]
test_dat <- BGGM::ptsd[201:221,]
fit_train <- estimate(train_dat, iter = 5000)


r2 <- predict(fit_train,
              measure = "R2",
              test_data  = test_dat,
              ci_width = 0.95,
              samples = 1000)

summary(r2)

plot(r2)

# plot test and training error
plot(x = error, x2 = r2)

}
\references{
Gelman, A., Goodrich, B., Gabry, J., & Vehtari, A. (2017). R-squared for Bayesian regression models.
            The American Statistician, (just-accepted), 1-6. \href{http://www.stat.columbia.edu/~gelman/research/unpublished/bayes_R2.pdf}{pre-print}
}
