% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.estimate.R
\name{predict.estimate}
\alias{predict.estimate}
\title{predictive accuracy}
\usage{
\method{predict}{estimate}(fit, test_data = NULL, ci_width,
  samples = 1000, measure = c("R2", "MSE"))
}
\arguments{
\item{fit}{fitted object of class estimate}

\item{test_data}{option test data set}

\item{ci_width}{width of the measure}

\item{samples}{number of samples used to}

\item{measure}{Bayesian R2 (R2) or mean squared error (MSE)}

\item{selected}{selected graph with credible intervals}
}
\value{
summary_error posterior mean, standard deviation, and credible interval for each variable

posterior_samples list containing the computed measure for each posterior draw
}
\description{
Bayesian predictive error. The measures are computed with respect to the posterior distributions.
This provides uncertaintly for variance explained (Bayesian R2) and mean squared error.
}
\examples{

dat <- BGGM::ptsd
fit <- estimate(dat, samples)

# select graph
selected <- select(fit, ci_width = 0.95)$adjacency_mat

error <- predict(fit,
              selected = selected,
              measure = "MSE",
              test_data  = NULL,
              ci_width = 0.95,
              samples = 1000)

error$summary_error

################################
## training/test data example ##
################################

train_dat <- BGGM::ptsd[1:200,]
fit_train <- estimate(train_dat, samples)

selected <- select(fit_train, ci_width = 0.95)$adjacency_mat

r2 <- predict(fit = fit_train,
              selected = selected,
              measure = "R2",
              test_data  = NULL,
              ci_width = 0.95,
              samples = 1000)

r2$summary_error
}
\references{
Gelman, A., Goodrich, B., Gabry, J., & Vehtari, A. (2017). R-squared for Bayesian regression models.
            The American Statistician, (just-accepted), 1-6. \href{http://www.stat.columbia.edu/~gelman/research/unpublished/bayes_R2.pdf}{pre-print}
}
