% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_BGGM.R
\name{print.BGGM}
\alias{print.BGGM}
\title{Print Method for \code{BGGM} Objects}
\usage{
\method{print}{BGGM}(x, ...)
}
\arguments{
\item{x}{An object of class \code{BGGM}.}

\item{...}{currently ignored}
}
\description{
Print Method for \code{BGGM} Objects
}
