% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{print.estimate}
\alias{print.estimate}
\title{Print method for \code{estimate.default} objects}
\usage{
\method{print}{estimate}(x, ...)
}
\arguments{
\item{x}{An object of class \code{estimate}}

\item{...}{currently ignored}
}
\description{
Print method for \code{estimate.default} objects
}
