% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{rmse}
\alias{rmse}
\title{Root Mean Squared Error}
\usage{
rmse(object, ...)
}
\arguments{
\item{object}{object of class \code{post.pred} or  \code{predict.estimate}}

\item{...}{currently ignored}
}
\value{

}
\description{
Root Mean Squared Error
}
\examples{
# data
Y <- subset(tas, gender == "M")[,-ncol(tas)]

# fit model
fit <- estimate(Y)

# predict (note summary = FALSE)
pred <- predict(fit, summary = FALSE)

rmse(pred)
}
