% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.estimate.R,
%   R/select.ggm_compare_estimate.R
\name{select.estimate}
\alias{select.estimate}
\alias{select.ggm_compare_estimate}
\title{Graph Selection for \code{estimate} Objects}
\usage{
\method{select}{estimate}(object, cred = 0.95, alternative = "two.sided", ...)

\method{select}{ggm_compare_estimate}(object, cred = 0.95, ...)
}
\arguments{
\item{object}{An object of class \code{estimate.default}.}

\item{cred}{Numeric. The credible interval width for selecting the graph
(defaults to 0.95; must be between 0 and 1).}

\item{alternative}{A character string specifying the alternative hypothesis. It
must be one of "two.sided" (default), "greater"  or "less".
See note for futher details.}

\item{...}{not currently used}
}
\value{
The returned object of class \code{select.estimate} contains a lot of information that
        is used for printing and plotting the results. For users of \strong{BGGM}, the following
        are the useful objects:

\itemize{

\item \code{pcor_adj} Selected partial correlation matrix (weighted adjacency).
\item \code{adj} Adjacency matrix for the selected edges
\item \code{object} An object of class \code{estimate} (the fitted model).

}

The returned object of class \code{select.ggm_compare_estimate} contains a lot of information that
        is used for printing and plotting the results. For users of \strong{BGGM}, the following
        are the useful objects:


\itemize{

\item \code{mean_diff} A list of matrices for each group comparsion (partial correlation differences).

\item \code{pcor_adj} A list of weighted adjacency matrices for each group comparsion.

\item \code{adj} A list of adjacency matrices for each group comparsion.

}
}
\description{
Provides the selected graph based on credible intervals for
the partial correlations that did not contain zero
\inserCite{Williams2019}{BGGM}.

Provides the selected graph (of differences) based on credible intervals for
the partial correlations that did not contain zero
\inserCite{Williams2019}{BGGM}.
}
\details{
This package was built for the social-behavioral sciences in particular. In these applications, there is
strong theory that expects \emph{all} effects to be positive. This is known as a "positive manifold" and
this notion has a rich tradition in psychometrics. Hence, this can be incorportated into the graph with
\code{alternative = "greater"}. This results in the estimted structure including only positive edges. Further
details can be found at the blog "Dealing with Negative (Red) Edges in Psychological Networks: Frequentist Edition"
(\href{https://donaldrwilliams.github.io/2020/03/29/dealing-with-negative-red-edges-in-psychological-networks-frequentist-edition/}{link})
}
\examples{
\donttest{
# data
Y <- bfi[,1:25]

# estimate
fit <- estimate(Y, iter = 250)


# select edge set
E <- select(fit)

}

\donttest{

##################
### example 1: ###
##################
data
Y <- bfi
males and females

Ymale <- subset(Y, gender == 1,
               select = -c(gender,
                           education))
# fit model
fit <- ggm_compare_estimate(Ymale, Yfemale,
                           type = "continuous")


E <- select(fit)

}
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{estimate}} and \code{\link{ggm_compare_estimate}} for several examples.
}
