% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.estimate.R
\name{select.estimate}
\alias{select.estimate}
\title{Title}
\usage{
\method{select}{estimate}(x, ci_width)
}
\arguments{
\item{x}{object of class \code{bayes_estimate}}

\item{ci_width}{credible interval width used for the decision rule}
}
\value{
the selected partial correlation matrix (the conditional dependence structure) and
 the adjacency matrix is returned
}
\description{
Title
}
