% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.explore.R
\name{select.explore}
\alias{select.explore}
\title{Title}
\usage{
\method{select}{explore}(x, BF_cut = 3, alternative = "two.sided",
  hyp_prob = NULL)
}
\arguments{
\item{x}{data matrix}

\item{BF_cut}{evidentiary threshold}

\item{alternative}{type of hypothesis (see notes)}

\item{hyp_prob}{posterior probability threshold (see notes)}
}
\description{
Title
}
