% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.explore.R
\name{select.explore}
\alias{select.explore}
\title{select explore}
\usage{
\method{select}{explore}(object, BF_cut = 3, alternative = "two.sided", ...)
}
\arguments{
\item{object}{object of class \code{explore.default}}

\item{BF_cut}{evidentiary threshold}

\item{alternative}{type of hypothesis (see notes)}

\item{...}{currently not used}
}
\value{
list of class \code{select.explore}:
}
\description{
select explore
}
