% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.estimate.R
\name{select}
\alias{select}
\title{S3 select method}
\usage{
select(x, ...)
}
\arguments{
\item{x}{object of class \code{estimate}}

\item{...}{not currently used}
}
\description{
S3 select method
}
\seealso{
\code{\link{select.estimate}}
}
