% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{summary.estimate}
\alias{summary.estimate}
\title{Summary method for \code{estimate.default} objects}
\usage{
\method{summary}{estimate}(object, ...)
}
\arguments{
\item{object}{An object of class \code{estimate}}

\item{...}{currently ignored}
}
\description{
Summary method for \code{estimate.default} objects
}
