% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.explore.R
\name{summary.select.explore}
\alias{summary.select.explore}
\title{Summary method for \code{select.explore} objects}
\usage{
\method{summary}{select.explore}(object, hyp = "H1", log = TRUE, summarize = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{select.explore}}

\item{hyp}{hypothesis to summarize (default is the alternative H1)}

\item{log}{log scale for the Bayes factors (default is \code{TRUE})}

\item{summarize}{if \code{TRUE} partial correlations and credible intervals are provided}

\item{...}{currently ignored}
}
\description{
Summary method for \code{select.explore} objects
}
\seealso{
\code{\link{select.explore}}
}
