% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.ggm_compare_bf.R
\name{summary.select.ggm_compare_bf}
\alias{summary.select.ggm_compare_bf}
\title{Print method for \code{select.ggm_compare_bf} objects}
\usage{
\method{summary}{select.ggm_compare_bf}(object, ...)
}
\arguments{
\item{object}{An object of class \code{select.ggm_compare_bf}}

\item{...}{currently ignored}
}
\description{
Print method for \code{select.ggm_compare_bf} objects
}
\seealso{
\code{\link{select.ggm_compare_bf}}
}
